% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\docType{package}
\name{AirSensor}
\alias{AirSensor}
\title{Data access and analysis functions for Purple Air sensor data}
\description{
This package contains code to access current synoptic data from
Purple Air as well as time series data for individual sensors from Thing
Speak.

Functions for downloading and enhancing sensor data return one of two types
of object:
\itemize{
\item{\code{pas} -- PurpleAirSynoptic dataframe of uniformly named properties}
\item{\code{pat} -- PurpleAirTimeseries list of dataframes containing
sensor metadata and data}
}

Analysis and visualization functions provide basic functionality for working
with Purple Air sensor data and comparing it with national monitoring data
retrieved with the \pkg{PWFSLSmoke} package.
}
