% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\docType{data}
\name{AIRSENSOR_1_PAT_FIELDS}
\alias{AIRSENSOR_1_PAT_FIELDS}
\title{Comma-separated list of fields needed to create AirSensor pat objects.}
\format{
String with comma-separated field names
}
\usage{
AIRSENSOR_1_PAT_FIELDS
}
\description{
Character string with default PurpleAir field names used in
\code{pat_createNew()}. These fields are sufficient for most
QC algorithms and include most of the "information and status" fields,
"humidity", "temperature", "pressure" and various PM "pseudo average" fields.
}
\references{
\href{https://api.purpleair.com/#api-sensors-get-sensor-history-csv}{Get Sensor History API}
}
\keyword{datasets}
