% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airSensorQC_AB_01.R
\name{airSensorQC_AB_01}
\alias{airSensorQC_AB_01}
\title{Apply QC to Aggregation Statistics}
\usage{
airSensorQC_AB_01(aggregationStats)
}
\arguments{
\item{aggregationStats}{Dataframe of statistics as returned by
\code{pat_aggregate()}.}
}
\value{
Data frame with columns \code{datetime} and \code{pm25}.
}
\description{
Creates a \code{pm25} timeseries by averaging aggregated data
from the A and B channels and applying the following QC logic:

\enumerate{
\item{Create pm25 by averaging the A and B channel aggregation means}
\item{Invalidate data where:  (min_count < 10)}
\item{Invalidate data where:  (p-value < 1e-4) & (mean_diff > 10)}
\item{Invalidate data where:  (pm25 < 100) & (mean_diff > 20)}
}
}
\examples{
\dontrun{
df <- 
  example_pat \%>\%
  pat_qc() \%>\%
  pat_aggregate() \%>\%
  airSensorQC_AB_01()
  
plot(df)
}
}
