% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\docType{data}
\name{ArchiveBaseDir}
\alias{ArchiveBaseDir}
\title{Base directory for pre-generated data}
\format{Directory string.}
\description{
If an archive of pre-generated data files is availalbe locally,
users can set the location of this directory with\code{setArchiveBaseDir()}.
Otherwise, users must specify an external source of pre-generated datafiles 
with \code{setArchiveBaseUrl()}.

To avoid internet latency, specification of BASE_DIR will always take 
precedence over specification of BASE_URL.

Package functions that load pre-generated data files will load data from this
directory. These functions include:

\itemize{
\item{\code{pas_load()}}
\item{\code{pat_load()}}
\item{\code{pat_loadLatest()}}
\item{\code{pat_loadMonth()}}
\item{\code{sensor_load()}}
\item{\code{sensor_loadLatest()}}
\item{\code{sensor_loadMonth()}}
}
}
\seealso{
getArchiveBaseDir

setArchiveBaseDir

setArchiveBaseUrl
}
\keyword{environment}
