% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_createAirSensor.R
\name{ast_createAirSensor}
\alias{ast_createAirSensor}
\title{Create a Air Sensor object}
\usage{
ast_createAirSensor(ast = NULL, parameter = "pm25")
}
\arguments{
\item{ast}{Air Sensor Timeseries object}

\item{parameter}{Which column of data from \code{ast} to use}
}
\value{
List with \code{meta} and \code{data} elements on a uniform time 
axis.

"as" list of aggregated time series PurpleAir data
}
\description{
Reformats an \code{ast} object into an "Air Sensor" object
that has appropriate metadata to be used with the *PWFSLSmoke* package.
}
\examples{
\dontrun{
pat <- 
  AirSensor::example_pat \%>\%
  pat_filterDate(20180701, 20180901)
ast <- pat_createASTimeseriesObject(pat, "1 hour")
as <- ast_createAirSensor(ast)
}
}
\seealso{
\link{pat_createASTimeseries}
}
