% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createASTimeseriesObject.R
\name{createASTimeseriesObject}
\alias{createASTimeseriesObject}
\title{Create a AirSensor Timeseries object}
\usage{
createASTimeseriesObject(pat, period = "15 min", ...)
}
\arguments{
\item{pat}{Purple Air timeseries data from \code{createPATimeseriesObject()}}

\item{period}{The time period to average to. Can be "sec", "min", "hour", 
"day", "DSTday", "week", "month", or "year". A number can also
precede these options followed by a space (i.e. "2 day" or "37 min").}

\item{...}{Extra parameters to fine tune time-averaging function}
}
\value{
List with original \code{meta} and restructured \code{data} elements

"as_timeseries" list of aggregated time series PurpleAir data
}
\description{
Aggregates a PurpleAir timeseries object along the datetime
axis to produce a new dataframe with 13 columns of period-aggregated
timeseries data: averages, standard deviations, measurement frequency, and 
qualitity control.
}
\examples{
\dontrun{
pat <- pat_load(pas, "Seattle", startdate = 20180701, enddate = 20180901)
seattle_AirSensor <- createASTimeseriesObject(pat, "2.5 hours")
}
}
\seealso{
\link{pat_timeAverage}
}
