% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPATimeseriesObject.R
\name{createPATimeseriesObject}
\alias{createPATimeseriesObject}
\title{Create a PurpleAir Timeseries object}
\usage{
createPATimeseriesObject(pat_raw = NULL)
}
\arguments{
\item{pat_raw}{Raw PurpleAir timeseries data from \code{pat_downloadParseData()}}
}
\value{
List with original \code{meta} and restructured \code{data} elements

"pa_timeseries" list of time series PurpleAir data
}
\description{
The 'data' dataframe is converted from 'long format' with temporal
resolution of seconds to 'wide format' with temporal resolution in minutes.
In the process, the following columns of data are omitted:
\itemize{
\item{\code{pm1_atm}}
\item{\code{pm2.5_atm}}
\item{\code{pm10_atm}}
}
}
\note{
On January 13, 2020 the PurpleAir FAQ "Whsat's the difference between CF_1
and CF_ATM?" contained the following text:

\preformatted{
The CF_ATM and CF_1 values are calculated from the particle count data with a 
proprietary algorithm developed by the PMS5003 laser counter manufacturer, 
PlanTower. The specifics of the calculation are not available to the public 
(or us for that matter). However, to convert the particle count data (um/dl) 
to a mass concentration (ug/m3) they must use an average particle density. 
They do provide 2 different mass concentration conversion options; CF_1 uses 
the "average particle density" for indoor particulate matter and CF_ATM uses 
the "average particle density" for outdoor particulate matter. Depending on 
the density of the particles you are measuring the sensor could appear to 
read "high" or "low". Some groups have developed conversion factors to 
convert the data from the sensor to match the unique average particle density 
within their airshed. 
}
}
\examples{
\dontrun{
library(AirSensor)
initializeMazamaSpatialUtils()

pat_raw <- pat_downloadParseData(
  label = 'North Bend Weather',
  pas = example_pas,
  startdate = 20180908
)

pat <- createPATimeseriesObject(pat_raw)
pat_multiplot(pat)
}
}
\references{
https://www2.purpleair.com/community/faq
}
\seealso{
\link{pat_downloadParseData}
}
