% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-temporalRange.R
\name{.dateRange}
\alias{.dateRange}
\title{Create a POSIXct date range}
\usage{
.dateRange(startdate = NULL, enddate = NULL, lookbackdays = 7,
  timezone = "UTC", unit = "sec")
}
\arguments{
\item{startdate}{desired start datetime (ISO 8601)}

\item{enddate}{desired end datetime (ISO 8601)}

\item{lookbackdays}{number of days of data to include}

\item{timezone}{Olson timezone used to interpret dates}

\item{unit}{units used to determine enddates final timestep}
}
\value{
A vector of two \code{POSIXct}s.
}
\description{
Uses incoming parameters to return a pair of \code{POSIXct}
times in the proper order. This function always rounds to day boundaries.

Dates can be anything that is understood by 
\code{lubrdiate::parse_date_time()} including either of the following 
recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}
}
\note{
The ending \code{POSIXct} will end one \code{unit} before the specified
\code{enddate}.
}
