% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-outliers.R
\name{.plotOutliers}
\alias{.plotOutliers}
\title{Plot flagged outliers}
\usage{
.plotOutliers(df = NULL, parameter = "pm25", ylim = NULL,
  xlab = "Date", ylab = "μg / m³", title = "default",
  subtitle = NULL, data_shape = 18, data_size = 1,
  data_color = "black", data_alpha = 0.5, outlier_shape = 8,
  outlier_size = 1, outlier_color = "red", outlier_alpha = 1)
}
\arguments{
\item{df}{data frame that contains datetime, PM2.5, and outlier boolean flags}

\item{parameter}{the parameter with the associated outliers flag}

\item{ylim}{y axis limits}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}

\item{subtitle}{plot substitle}

\item{data_shape}{shape of points}

\item{data_size}{size of points}

\item{data_color}{color of points}

\item{data_alpha}{alpha of points}

\item{outlier_shape}{outlier point shape}

\item{outlier_size}{outlier point size}

\item{outlier_color}{outlier point color}

\item{outlier_alpha}{outlier point alpha}
}
\value{
gg object
}
\description{
An internal capability to quickly plot outlier data.
}
