% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gen.R
\name{.sample}
\alias{.sample}
\title{General table row sampling}
\usage{
.sample(data, sampleSize = NULL, sampleFraction = 1, setSeed = NULL)
}
\arguments{
\item{data}{Dataframe to be sampled}

\item{sampleSize}{a non-negative integer giving the number of rows to choose.}

\item{sampleFraction}{the fraction of rows to choose.
\code{smapleSize} is uised)}

\item{setSeed}{an integer that sets random numbver generation. Can be used to 
reproduce sampling.}
}
\value{
A data.frame
}
\description{
This is a wrapper around sample() to make it easy to select 
random rows from a table. Supports both integer (sampleSize) and fractional 
(sampleFraction) N row sampling. For reproducible debugging, set a hash seed.
}
\keyword{internal}
