% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadParseSensorList.R
\name{downloadParseSensorList}
\alias{downloadParseSensorList}
\title{Download Sensor list from PurpleAir}
\usage{

  downloadParseSensorList(baseUrl = "https://www.purpleair.com/sensorlist")
}
\arguments{
\item{baseUrl}{base URL for Sensor archive}
}
\value{
A data.frame of PurpleAir sensors (inactive or active)
}
\description{
Download a complete list of the PurpleAir sensor archive. 
Thoeretically includes deprectaed/inactive sensors. The list can be manually
accessed via web browser provided in the reference.
}
\examples{
\dontrun{
initializeMazamaSpatialUtils()
sensor_list <- downloadParseSensorList()
}
}
\references{
\href{https://www.purpleair.com/sensorlist}{PurpleAir Sensor List}
}
\seealso{
\link{downloadParseSynopticData}
}
