% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadParseSynopticData.R
\name{downloadParseSynopticData}
\alias{downloadParseSynopticData}
\title{Download synoptic data from Purple Air}
\usage{
downloadParseSynopticData(baseUrl = "https://www.purpleair.com/json")
}
\arguments{
\item{baseUrl}{base URL for synoptic data}
}
\value{
Dataframe of synoptic PurpleAir data.
}
\description{
Download and parse synoptic data from the Purple Air network
of particulate monitors.

The synoptic data provides a view of the entire Purple Air network and
includes both metadata and recent PM2.5 averages for each deployed monitor.
}
\examples{
\dontrun{
initializeMazamaSpatialUtils()
pas_raw <- downloadParseSynopticData()
View(pas_raw[1:100,])
}
}
\references{
\href{https://www.purpleair.com/json}{json formatted Purple Air data}
}
\seealso{
\link{enhanceSynopticData}
}
