% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\docType{data}
\name{example_pas_raw}
\alias{example_pas_raw}
\title{Example raw Purple Air Synoptic dataset}
\format{A tibble with 12657 rows and 32 columns of data.}
\source{
https://www.purpleair.com/json
}
\description{
The \code{example_pas_raw} dataset provides a quickly loadable
version of raw Purple Air synoptic data JSON for practicing and code
examples. This dataset was was generated on 2019-07-06 by running:

\preformatted{
  example_pas_raw <- downloadParseSynopticData()
}

This dataset can be converted into a standard \emph{pas} dataset with:

\preformatted{
  pas <- enhanceSynopticData(example_pas_raw)
}
}
\seealso{
example_pas
}
\keyword{datasets}
