% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\docType{data}
\name{example_pat_failure_B}
\alias{example_pat_failure_B}
\title{Example Purple Air Timeseries dataset exhibiting severe errors}
\format{An S3 object composed of "meta" and "data" data.}
\description{
The example_pat_failure_b dataset provides a quickly loadable
version of a \emph{pa_timeseries} object for practicing and code development. 
This dataset was generated by running \code{pat_loadLatest()} for the SCTV_16
sensor in Los Angeles over a two-week interval, 2019-06-01 to 2019-06-18.
}
\seealso{
createPATimeseriesObject
}
\keyword{datasets}
