% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\docType{data}
\name{example_sensor}
\alias{example_sensor}
\title{Example AirSensor Timeseries dataset}
\format{An S3 object composed of "meta" and "data" data.}
\description{
The example_as dataset provides a quickly loadable version of
a \emph{airsensor} object for practicing and code development. This dataset 
was generated by running \code{pat_createAirSensor} for on a monitor
from a timeseries generated by 
\code{
pat_load("SCAN_14", "20180801", "20181001")
}
}
\seealso{
pat_loadLatest
}
\keyword{datasets}
