% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_sensor_scaqmd}
\alias{example_sensor_scaqmd}
\title{Example AirSensor Timeseries dataset}
\format{
An S3 object composed of "meta" and "data" data.
}
\usage{
example_sensor_scaqmd
}
\description{
The \code{example_sensor_scaqmd} dataset provides a quickly
loadable version of a multi-sensor \emph{airsensor} object for practicing and
code examples. This dataset was was generated on 2020-09-15 by running:

\preformatted{
library(AirSensor)

setArchiveBaseUrl("https://airsensor.aqmd.gov/PurpleAir/v1")

example_sensor_scaqmd <-
  sensor_load("scaqmd", startdate = 20190701, enddate = 20190708)

save(example_sensor_scaqmd, file = "data/example_sensor_scaqmd.rda")
}
}
\keyword{datasets}
