% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gen.R
\name{flagOutliers}
\alias{flagOutliers}
\title{Flag outliers in vectorized data}
\usage{
flagOutliers(df, parameter = NULL, windowSize = 23, thresholdMin = 8)
}
\arguments{
\item{df}{A data frame.}

\item{parameter}{The data frame parameter to check and replace outliers}

\item{windowSize}{The size of the rolling window}

\item{thresholdMin}{The minimum threshold value to detect outliers}
}
\value{
\code{data.frame} A data.frame with an additional flag vector
}
\description{
Outlier detection using Hampel identification.  For each sample 
 of a vector, the function computes the median of a center-aligned window composed of the sample 
 and its \code{windowSize} surrounding samples. It also estimates the 
 standard deviation of each sample about its window median using the median 
 absolute deviation. If a sample differs from the median by more than three 
 standard deviations, it is replaced with the median. If x is a matrix, then 
 hampel treats each column of x as an independent channel.
 
 The \code{thresholdMin} level is similar to a sigma value for normally 
 distributed data.
}
