% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\name{getArchiveBaseDir}
\alias{getArchiveBaseDir}
\title{Get data archive base directory}
\usage{
getArchiveBaseDir()
}
\value{
directory string.
}
\description{
Returns the package base directory pointing to an archive of
pre-generated data files.
}
\seealso{
archiveBaseDir

setArchiveBaseDir
}
\keyword{environment}
