% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\name{getArchiveBaseUrl}
\alias{getArchiveBaseUrl}
\title{Get data archive base URL}
\usage{
getArchiveBaseUrl()
}
\value{
URL string.
}
\description{
Returns the package base URL pointing to an archive of
pre-generated data files.
}
\seealso{
archiveBaseUrl

setArchiveBaseUrl
}
\keyword{environment}
