% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeMazamaSpatialUtils.R
\name{initializeMazamaSpatialUtils}
\alias{initializeMazamaSpatialUtils}
\title{Initialize Mazama Spatial Utils}
\usage{
initializeMazamaSpatialUtils(spatialDataDir = "~/Data/Spatial",
  stateCodeDataset = "NaturalEarthAdm1", logLevel = WARN)
}
\arguments{
\item{spatialDataDir}{directory where spatial datasets are created}

\item{stateCodeDataset}{MazamaSpatialUtils dataset returning ISO 3166-2 
alpha-2 stateCodes}

\item{logLevel}{directory where spatial datasets are created}
}
\description{
Convenience function that wraps:

\preformatted{
  logger.setup()
  logger.setLevel(WARN)
  MazamaSpatialUtils::setSpatialDataDir('~/Data/Spatial')
  MazamaSpatialUtils::loadSpatialData('NaturalEarthAdm1')
}

If file logging is desired, these commands should be run individually with
output log files specified as arguments to \code{logger.setup()} from the
\pkg{MazamaCoreUtils} package.
}
