% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFile.R
\name{loadDataFile}
\alias{loadDataFile}
\title{Load data from URL or local file}
\usage{
loadDataFile(filename = NULL, baseUrl = NULL, dataDir = NULL)
}
\arguments{
\item{filename}{Name of the data file to be loaded.}

\item{baseUrl}{Base URL for data files.}

\item{dataDir}{Local directory containing data files.}
}
\value{
A data object.
}
\description{
Loads pre-generated .rda RData files from a URL or a local
directory. This function is intended to be called by other \code{~_load()}
functions and can remove internet latencies when local versions of daata are
available.

For this reason, specification of \code{dataDir} always takes precedence over 
\code{baseUrl}.
}
\keyword{internal}
