% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\name{pa_createMember}
\alias{pa_createMember}
\title{Create a new member within the specified group.}
\usage{
pa_createMember(
  api_key = NULL,
  group_id = NULL,
  sensor_index = NULL,
  baseUrl = "https://api.purpleair.com/#api-groups-create-member"
)
}
\arguments{
\item{api_key}{PurpleAir API WRITE key.}

\item{group_id}{The \code{group_id} of the requested group. This group must
be owned by the \code{api_key}.}

\item{sensor_index}{Sensor index as returned by \code{pa_getSensorsData()}.}

\item{baseUrl}{URL endpoint for the "Create Member" API.}
}
\value{
List containing data associated with this this sensor.
}
\description{
Sends a request to the PurpleAir API endpoint described at:
\url{https://api.purpleair.com/#api-groups-create-member}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

  pa_getGroupDetail(
    api_key = PURPLE_AIR_API_READ_KEY,
    group_id = MY_GROUP_ID,
    sensor_index = MY_SENSOR_INDEX
  )

}, silent = FALSE)
}
}
