% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\name{pa_deleteGroup}
\alias{pa_deleteGroup}
\title{Delete the specified group.}
\usage{
pa_deleteGroup(
  api_key = NULL,
  group_id = NULL,
  baseUrl = "https://api.purpleair.com/v1/groups"
)
}
\arguments{
\item{api_key}{PurpleAir API WITE key.}

\item{group_id}{The \code{group_id} to be deleted. This group must
be owned by the \code{api_key}.}

\item{baseUrl}{URL endpoint for the "Delete Group" API.}
}
\value{
No return.
}
\description{
Sends a request to the PurpleAirAPI API endpoint described at:
\url{https://api.purpleair.com/#api-groups-delete-group}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

  pa_deleteGroup(
    api_key = PURPLE_AIR_API_READ_KEY,
    group_id = MY_GROUP_ID
  )

}, silent = FALSE)
}
}
