% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\name{pa_deleteMember}
\alias{pa_deleteMember}
\title{Delete a member from the specified group.}
\usage{
pa_deleteMember(
  api_key = NULL,
  group_id = NULL,
  member_id = NULL,
  baseUrl = "https://api.purpleair.com/v1/groups"
)
}
\arguments{
\item{api_key}{PurpleAir API WITE key.}

\item{group_id}{The \code{group_id} of the requested group.}

\item{member_id}{The \code{member_id} to be deleted.}

\item{baseUrl}{URL endpoint for the "Delete Member" API.}
}
\value{
No return.
}
\description{
Sends a request to the PurpleAirAPI API endpoint described at:
\url{https://api.purpleair.com/#api-groups-delete-member}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

  pa_getGroupDetail(
    api_key = MY_API_READ_KEY,
    group_id = MY_GROUP_ID
  )

}, silent = FALSE)
}
}
