% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\name{pa_getGroupDetail}
\alias{pa_getGroupDetail}
\title{Retrieve all members of the specified group.}
\usage{
pa_getGroupDetail(
  api_key = NULL,
  group_id = NULL,
  baseUrl = "https://api.purpleair.com/v1/groups"
)
}
\arguments{
\item{api_key}{PurpleAir API READ key.}

\item{group_id}{The \code{group_id} of the requested group. This group must
be owned by the \code{api_key}.}

\item{baseUrl}{URL endpoint for the "Get Group Detail" API.}
}
\value{
List containing all members of the specified group.
}
\description{
Sends a request to the PurpleAirAPI API endpoint described at:
\url{https://api.purpleair.com/#api-groups-get-group-detail}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

  pa_getGroupDetail(
    api_key = MY_API_READ_KEY,
    group_id = MY_GROUP_ID
  )

}, silent = FALSE)
}
}
