% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\name{pa_getGroupsList}
\alias{pa_getGroupsList}
\title{Retrieve all groups owned by the provided \code{api_key}.}
\usage{
pa_getGroupsList(
  api_key = NULL,
  baseUrl = "https://api.purpleair.com/v1/groups"
)
}
\arguments{
\item{api_key}{PurpleAir API READ ey.}

\item{baseUrl}{URL endpoint for the "Get Groups List" API.}
}
\value{
List containing all groups owned by \code{api_key}.
}
\description{
Sends a request to the PurpleAirAPI API endpoint described at:
\url{https://api.purpleair.com/#api-groups-get-groups-list}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

  pa_getGroupsList(
    api_key = PURPLE_AIR_API_READ_KEY
  )

}, silent = FALSE)
}
}
