% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\name{pa_getMemberData}
\alias{pa_getMemberData}
\title{Retrieve recent data for a single sensor in the specified group.}
\usage{
pa_getMemberData(
  api_key = NULL,
  group_id = NULL,
  member_id = NULL,
  fields = NULL,
  baseUrl = "https://api.purpleair.com/v1/groups"
)
}
\arguments{
\item{api_key}{PurpleAir API READ key.}

\item{group_id}{The \code{group_id} of the requested group. This group must
be owned by the \code{api_key}.}

\item{member_id}{Unique \code{member_id} for a sensor within \code{group_id}.}

\item{fields}{Optional parameter specifying sensor data fields to return.}

\item{baseUrl}{URL endpoint for the "Get Member Data" API.}
}
\value{
List containing all recent data for a single sensor.
}
\description{
Sends a request to the PurpleAir API endpoint described at:
\url{https://api.purpleair.com/#api-groups-get-member-data}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

  pa_getMemberData(
    api_key = MY_API_READ_KEY,
    group_id = MY_GROUP_ID,
    member_id = MY_MEMBER_ID
  )

}, silent = FALSE)
}
}
