% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\name{pa_getMemberHistory}
\alias{pa_getMemberHistory}
\title{Retrieve historical data for a single sensor of the specified group.}
\usage{
pa_getMemberHistory(
  api_key = NULL,
  group_id = NULL,
  member_id = NULL,
  start_timestamp = NULL,
  end_timestamp = NULL,
  average = 10,
  fields = SENSOR_HISTORY_PM25_FIELDS,
  baseUrl = "https://api.purpleair.com/v1/groups"
)
}
\arguments{
\item{api_key}{PurpleAir API READ key.}

\item{group_id}{The \code{group_id} of the requested group. This group must
be owned by the \code{api_key}.}

\item{member_id}{Unique \code{member_id} for a sensor within \code{group_id}.}

\item{start_timestamp}{Desired start datetime (ISO 8601).}

\item{end_timestamp}{Desired end datetime (ISO 8601).}

\item{average}{Temporal averaging in minutes performed by PurpleAir. One of:
0 (raw), 10, 30, 60 (hour), 360, 1440 (day).}

\item{fields}{Character string specifying which 'sensor data fields' to include in the response.}

\item{baseUrl}{URL endpoint for the "Get Groups list" API.}
}
\value{
Tibble with historical data for a single sensor.
}
\description{
Sends a request to the PurpleAirAPI API endpoint described at:
\url{https://api.purpleair.com/#api-groups-get-member-history}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

  pa_getMemberData(
    api_key = PURPLE_AIR_API_READ_KEY,
    group_id = MY_GROUP_ID,
    member_id = MY_MEMBER_ID
  )

}, silent = FALSE)
}
}
