% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\name{pa_getMembersData}
\alias{pa_getMembersData}
\title{Retrieve current data for all sensors in the specified group.}
\usage{
pa_getMembersData(
  api_key = NULL,
  group_id = NULL,
  fields = SENSOR_DATA_PM25_FIELDS,
  location_type = NULL,
  max_age = 604800,
  baseUrl = "https://api.purpleair.com/v1/groups"
)
}
\arguments{
\item{api_key}{PurpleAir API READ key.}

\item{group_id}{The \code{group_id} of the requested group. This group must
be owned by the \code{api_key}.}

\item{fields}{Comma-separated list of 'sensor data fields' to include in the response.}

\item{location_type}{The \code{location_type} of the sensors. Possible values
are: 0 = Outside, 1 = Inside or \code{NULL} = both.}

\item{max_age}{Filter results to only include sensors modified or updated
within the last \code{max_age} seconds. Using a value of 0 will match sensors of any age.}

\item{baseUrl}{URL endpoint for the "Get Members Data" API.}
}
\value{
List containing current data for all sensors in the specified group.
}
\description{
Sends a request to the PurpleAir API endpoint described at:
\url{https://api.purpleair.com/#api-groups-get-members-data}

Retrieves data for all sensors in the specified group.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

  pa_getMembersData(
    api_key = PURPLE_AIR_API_READ_KEY,
    group_id = MY_GROUP_ID
  )

}, silent = FALSE)
}
}
