% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\name{pa_getSensorData}
\alias{pa_getSensorData}
\title{Retrieve the latest data of a single sensor.}
\usage{
pa_getSensorData(
  api_key = NULL,
  sensor_index = NULL,
  fields = SENSOR_DATA_PM25_FIELDS,
  baseUrl = "https://api.purpleair.com/v1/sensors"
)
}
\arguments{
\item{api_key}{PurpleAir API READ key.}

\item{sensor_index}{The \code{sensor_index} as found in the JSON for this
specific sensor.}

\item{fields}{Optional parameter specifying sensor data fields to return.}

\item{baseUrl}{URL endpoint for the "Get Member Data" API.}
}
\value{
List containing all recent data for a single sensor.
}
\description{
Sends a request to the PurpleAir API endpoint described at:
\url{https://api.purpleair.com/#api-sensors-get-sensor-data}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

  pa_getSensorData(
    api_key = PURPLE_AIR_API_READ_KEY,
    sensor_index = MY_SENSOR_INDEX,
    fields = SENSOR_DATA_PM25_FIELDS
  )

}, silent = FALSE)
}
}
