% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\name{pa_getSensorHistory}
\alias{pa_getSensorHistory}
\title{Retrieve historical data for a single sensor.}
\usage{
pa_getSensorHistory(
  api_key = NULL,
  sensor_index = NULL,
  start_timestamp = NULL,
  end_timestamp = NULL,
  average = 10,
  fields = SENSOR_HISTORY_PM25_FIELDS,
  baseUrl = "https://api.purpleair.com/v1/sensors"
)
}
\arguments{
\item{api_key}{PurpleAir API READ key.}

\item{sensor_index}{The \code{sensor_index} as found in the JSON for this
specific sensor.}

\item{start_timestamp}{Optional Unix timestamp in seconds since Jan 1, 1970.}

\item{end_timestamp}{Optional Unix timestamp in seconds since Jan 1, 1970.}

\item{average}{Temporal averaging in minutes performed by PurpleAir. One of:
0 (raw), 10, 30, 60 (hour), 360, 1440 (day).}

\item{fields}{Character string specifying which 'sensor data fields' to include in the response.}

\item{baseUrl}{URL endpoint for the "Get Sensor History" API.}
}
\value{
List with historical data for a single sensor.
}
\description{
Sends a request to the PurpleAir API endpoint described at:
\url{https://api.purpleair.com/#api-sensors-get-sensor-history}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

start <-
  MazamaCoreUtils::parseDatetime("2023-01-29 00:00:00", timezone = "UTC") \%>\%
  as.numeric()

end <-
  MazamaCoreUtils::parseDatetime("2023-01-30 00:00:00", timezone = "UTC") \%>\%
  as.numeric()

pa_getSensorHistory(
  api_key = MY_API_READ_KEY,
  sensor_index = 896,
  start_timestamp = start,
  end_timestamp = end,
  average = 0,
  fields = SENSOR_HISTORY_PM25_FIELDS
)

}, silent = FALSE)
}
}
