% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas.R, R/pas_Jan25.R
\docType{data}
\name{pas}
\alias{pas}
\title{Synoptic Purple Air sensor data, enhanced}
\format{a dataframe (also a pas_synoptic object) with 6967 observations of 
35 variables.}
\usage{
pas

pas
}
\description{
Purple Air Sensor data returned by pas_load, with human readable column names.

Purple Air Sensor data returned by pas_load, with human readable column names.
}
\keyword{datasets}
