% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_addCommunityRegion.R
\name{pas_addCommunityRegion}
\alias{pas_addCommunityRegion}
\title{Add an air district to PurpleAir Synoptic Data}
\usage{
pas_addCommunityRegion(pas = NULL)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}
}
\value{
A pa_synoptic dataframe
}
\description{
Adds an air district (if any) to a pa_synoptic object via the 
\code{MazamaSpatialUtils} Package using PurpleAir location coordinates to 
determine the air basin the sensor is in.
}
\note{
As of 2020-04-14, only California air basins is supported.
}
\examples{
\dontrun{
example_pas_raw \%>\% 
  pas_addSpatialMetadata() \%>\% 
  pas_addCommunityRegion()
}
}
\seealso{
\link{pas_enhanceData}
}
