% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_downloadParseRawData.R
\name{pas_downloadParseRawData}
\alias{pas_downloadParseRawData}
\title{Download synoptic data from PurpleAir}
\usage{
pas_downloadParseRawData(
  apiReadKey = NULL,
  maxAge = 3600 * 24 * 7,
  outsideOnly = TRUE,
  west = NULL,
  east = NULL,
  south = NULL,
  north = NULL,
  baseUrl = "https://api.purpleair.com/v1/sensors"
)
}
\arguments{
\item{apiReadKey}{PurpleAir API Read Key. If \code{apiReadKey = NULL}, it
will be obtained using \code{getAPIKey("PurpleAir-read")}.
See \code{MazamaCoreUtils::\link[MazamaCoreUtils:setAPIKey]{setAPIKey}}.}

\item{maxAge}{Number of seconds used to filter results to only include sensors
modified or updated within the \code{maxAge} seconds. Using a value of 0 will match all sensors.}

\item{outsideOnly}{Logical specifying whether to restrict requests to outside sensors only.}

\item{west}{Longitude of the western edge of the bounding box in which to find sensors.}

\item{east}{Longitude of the eastern edge of the bounding box in which to find sensors.}

\item{south}{Latitude of the southern edge of the bounding box in which to find sensors.}

\item{north}{Latitude of the northern edge of the bounding box in which to find sensors.}

\item{baseUrl}{Base URL for the PurpleAir API.}
}
\value{
Dataframe of synoptic PurpleAir data.
}
\description{
Download and parse synoptic data for PurpleAir within the
specified region.

The synoptic data provides access to data from many PurpleAir sensors at
a moment in time and includes both metadata and recent PM2.5 averages for
each sensor.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

pas_raw <-
  pas_downloadParseRawData(
    apiReadKey = API_READ_KEY,
    maxAge = 3600 * 24,
    outsideOnly = TRUE,
    west = -125,
    east = -117,
    south = 42,
    north = 49
  )

View(pas_raw[1:100,])

}, silent = FALSE)
}
}
\references{
\href{https://www2.purpleair.com}{PurpleAir}

\href{https://api.purpleair.com}{PurpleAir API}

\href{https://www2.purpleair.com/policies/terms-of-service}{PurpleAir Terms of service}

\href{https://www2.purpleair.com/pages/license}{PurpleAir Data license}

\href{https://www2.purpleair.com/pages/attribution}{PurpleAir Data Attribution}
}
