% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_esriMap.R
\name{pas_esriMap}
\alias{pas_esriMap}
\title{ESRI static map of Purple Air sensors}
\usage{
pas_esriMap(pas, centerLon = NULL, centerLat = NULL,
  maptype = "worldStreetMap", mapRaster = NULL, width = 800,
  height = 800, zoom = NULL, shape = 15, size = 4, alpha = 0.8,
  theme = "Purples", scaleMin = 0, scaleMax = 150)
}
\arguments{
\item{pas}{\emph{pa_synoptic} object}

\item{centerLon}{map center longitude}

\item{centerLat}{map center latitude}

\item{maptype}{map type}

\item{mapRaster}{optional RGB Raster object to be used as the map layer}

\item{width}{width of image in pixels}

\item{height}{height of image in pixels}

\item{zoom}{map zoom level}

\item{shape}{symbol to use for points}

\item{size}{size of points}

\item{alpha}{opacity of points}

\item{theme}{base color or palette to be used.}

\item{scaleMin}{default is 0, minimum value to set scale for color gradient}

\item{scaleMax}{default is 150, maximum value to set scale for color gradient}
}
\value{
Plots a map loaded from arcGIS REST with points for each monitor.
}
\description{
Creates an ESRI map of a \emph{pa_synoptic} object.

Available \code{maptype} options include:
\itemize{
\item{natGeo}
\item{worldStreetMap}
\item{worldTopoMap}
\item{satellite}
\item{deLorme}
}

Available \code{theme} options include sequential and diverging palettes:

The sequential palettes names are
 
\code{"Blues" "BuGn" "BuPu" "GnBu" "Greens" "Greys" "Oranges" "OrRd" "PuBu" 
"PuBuGn" "PuRd" "Purples" "RdPu" "Reds" "YlGn" "YlGnBu" "YlOrBr" "YlOrRd"}

The diverging palettes are 

\code{"BrBG" "PiYG" "PRGn" "PuOr" "RdBu" "RdGy" "RdYlBu" "RdYlGn" "Spectral"}


Additional base maps are found at:
\url{https://developers.arcgis.com/}

If \code{centerLon}, \code{centerMap} or \code{zoom} are not specified,
appropriate values will be calcualted using data from the
\code{pa_synoptic} dataframe.
}
\examples{
\dontrun{
ca <- pas_load() \%>\% filter(stateCode == "CA")
pas_esriMap(ca)
}
\dontrun{
pas \%>\%
  pas_filter(stateCode == "CA" & pm25 > 75) \%>\%
  pas_esriMap(theme = "Purples", scaleMax = 250)
}

}
\seealso{
\code{\link{esriMap_plotOnStaticMap}}
}
\keyword{pa_synoptic}
