% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_filter.R
\name{pas_filter}
\alias{pas_filter}
\title{General purpose PurpleAir synoptic filter}
\usage{
pas_filter(pas, ...)
}
\arguments{
\item{pas}{a pa_synoptic dataframe}

\item{...}{Logical predicates defined in terms of the variables in the 
\code{pas}. Multiple conditions are combined with & or seperated by a comma. 
Only rows where the condition evaluates to TRUE are kept.}
}
\value{
a pa_synoptic object
}
\description{
A generalized data filter for pa_synoptic \emph{pas} objects to 
choose rows/cases where conditions are true. Rows where the condition 
evaluates to NA are dropped.
}
\examples{
\dontrun{
californiaSensors <- pas_filter(pas, stateCode == "CA") 
}
}
\seealso{
\link{pas_load}
}
