% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_filterNear.R
\name{pas_filterNear}
\alias{pas_filterNear}
\title{Find Purple Air sensors within radial distance}
\usage{
pas_filterNear(pas, latitude = NULL, longitude = NULL,
  radius = "1 km")
}
\arguments{
\item{pas}{a \code{pa_synoptic} object.}

\item{latitude}{a target latitude.}

\item{longitude}{a target longitude.}

\item{radius}{a distance with unit (i.e "15 km").}
}
\value{
A subset of the given \code{pa_synoptic} object.
}
\description{
Filter for Purple Air sensors within a specified distance from 
specified target coordinates.
}
\details{
\code{radius} Should be a numeric string with provided metric unit 
separated by a space, such as \code{"250 m"}.
}
\examples{
\dontrun{
pas_filterNear(
    pas = example_pas, 
    label = "SaratogaMiddleHighSchool", 
    radius = "1024 km"
    )
}
\dontrun{
pas_filterNear(
    pas = example_pas,
    latitude = 47.61702, 
    longitude = -122.34376
    )
}

}
\seealso{
\link{pas_filter}, \link{pas_filterArea}
}
