% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_getIDs.R
\name{pas_getIDs}
\alias{pas_getIDs}
\title{Return ID column of data from filtered PurpleAir Synoptic objects}
\usage{
pas_getIDs(pas = NULL)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}
}
\value{
A column of data.
}
\description{
A filter for \emph{pas} objects to return the ID column of the 
\emph{pas} object.
}
