% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paSynoptic.R
\name{pas_isPas}
\alias{pas_isPas}
\title{Test for correct structure in a \emph{pas} object}
\usage{
pas_isPas(pas)
}
\arguments{
\item{pas}{\emph{pas} object}
}
\value{
\code{TRUE} if \code{pas} has the correct structure, \code{FALSE} otherwise.
}
\description{
Test for correct structure in a \emph{pas} object
}
\keyword{pas}
