% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shiny.R
\name{pas_leaflet_shiny}
\alias{pas_leaflet_shiny}
\title{Leaflet interactive map for use with AirShiny}
\usage{
pas_leaflet_shiny(pas = NULL, parameter = "pm25_1hr",
  paletteName = NULL, radius = 11, opacity = 0.8,
  maptype = "terrain")
}
\arguments{
\item{pas}{Enhanced dataframe of PurpleAir synoptic data.}

\item{parameter}{Value to plot, e.g. \code{pm25_1hr}.}

\item{paletteName}{A predefined color palette name.}

\item{radius}{Radius (pixels) of monitor circles.}

\item{opacity}{Opacity of monitor circles.}

\item{maptype}{Optional name of leaflet ProviderTiles to use, e.g. \code{terrain}.}
}
\value{
A leaflet "plot" object which, if not assigned, is rendered in 
Rstudio's 'Viewer' tab.
}
\description{
This function creates interactive maps that will be displayed in 
AirShiny web-app.
}
\details{
The \code{maptype} argument is mapped onto leaflet "ProviderTile" 
names. Current mappings include:
\enumerate{
\item{"roadmap"}{ -- "OpenStreetMap"}
\item{"satellite"}{ -- "Esri.WorldImagery"}
\item{"terrain"}{ -- "Esri.WorldTopoMap"}
\item{"toner"}{ -- "Stamen.Toner"}
}

If a character string not listed above is provided, it will be used as the 
underlying map tile if available. See 
\url{https://leaflet-extras.github.io/leaflet-providers/} for a list of 
"provider tiles" to use as the background map.
}
