% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_load.R
\name{pas_load}
\alias{pas_load}
\title{Get latest Purple Air synoptic data}
\usage{
pas_load(baseUrl = "https://www.purpleair.com/json",
  countryCodes = c("US"), includePWFSL = TRUE, lookbackDays = 1)
}
\arguments{
\item{baseUrl}{Base URL for synoptic data.}

\item{countryCodes}{ISO country codes used to subset the data.}

\item{includePWFSL}{Logical specifying whether to calculate distances from 
PWFSL monitors.}

\item{lookbackDays}{Number of days to "look back" for valid data. Data are
filtered to only include sensors with data more recent than 
\code{lookbackDays} ago.}
}
\value{
Enhanced dataframe of synoptic Purple Air data.
}
\description{
Download, parse and enhance synoptic data from PurpleAir and
return the results as auseful tibble with class \code{pa_synoptic}.

Steps include:

1) Download and parse synoptic data

2) Replace variable with more consistent, more human readable names.

3) Add spatial metadata for each monitor including:
\itemize{
  \item{timezone -- olson timezone}
  \item{countryCode -- ISO 3166-1 alpha-2}
  \item{stateCode -- ISO 3166-2 alpha-2}
}

4) Convert data types from character to \code{POSIXct} and \code{numeric}.

5) Add distance and monitorID for the closest PWFSL monitor

Subsetting by country may be performed by specifying the \code{countryCodes}
argument.
}
\examples{
\dontrun{
initializeMazamaSpatialUtils()
pas <- pas_load()
}
}
\seealso{
\link{downloadParseSynopticData}
}
