% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_raw.R
\docType{data}
\name{pas_raw}
\alias{pas_raw}
\title{Synoptic Purple Air Sensor data, pre-enhancement"}
\format{A dataframe with 11663 observations of 32 variables.}
\source{
https://www.purpleair.com/json
}
\usage{
pas_raw
}
\description{
Purple Air Sensor data returned by downloadParseSynopticData(), with some 
human un-readable column names.
}
\keyword{datasets}
