% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_sensorDeploiymentID.R
\name{pas_sensorDeploymentID}
\alias{pas_sensorDeploymentID}
\title{Create sensor-deployent IDs for specific sensors}
\usage{
pas_sensorDeploymentID(pas = NULL, ids = NULL)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{ids}{Vector of sensor IDs.}
}
\value{
Vector of time series identifiers.
}
\description{
The "sensor-deployment" ID is the unique time series identifier
used in \emph{pat} archive file names and in \emph{airsensor} objects. It is
created by combining a digest of the location (with an accuracy of < 0.1 
meters) with the device ID.

It is intended to be associated with the parent ID (A channel)
in the \code{pas} object.
}
