% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_staticMap.R
\name{pas_staticMap}
\alias{pas_staticMap}
\title{Static map of Purple Air sensors}
\usage{
pas_staticMap(pas, parameter = "pm25_1hr", palette = "Purples",
  mapTheme = "terrain", mapShape = "sq", direction = 1,
  minScale = 0, maxScale = 150, shape = 15, size = 2,
  alpha = 0.8, bbuff = 0.1, zoomAdjust = 0, ...)
}
\arguments{
\item{pas}{\emph{pa_synoptic} object}

\item{parameter}{Value to plot, e.g. \code{pm25_1hr}.}

\item{palette}{base color or palette to be used.}

\item{mapTheme}{default is "terrain", see description for additional options}

\item{mapShape}{default is "square", can also be "natural"}

\item{direction}{legend color direction}

\item{minScale}{default is 0, minimum value to set scale for color gradient}

\item{maxScale}{default is 150, maximum value to set scale for color gradient}

\item{shape}{symbol to use for points}

\item{size}{size of points}

\item{alpha}{opacity of points}

\item{bbuff}{bounding box buffer, default is 0.1}

\item{zoomAdjust}{adjustment to map zoom level (-1:3)}

\item{...}{additional options: the legend can disabled \code{guide = FALSE}}
}
\value{
ggplot object
}
\description{
Creates a static map of a \emph{pa_synoptic} object.

Users can create a map using any numeric data column within the
\emph{pa_synoptic} object:

\code{"pm25" "temperature" "humidity" "pressure" 
 "pm25_current" "pm25_10min" "pm25_30min" "pm25_1hr" "pm25_6hr" 
 "pm25_1day" "pm25_1week" "pwfsl_closestDistance"}

Available \code{palette} options include an \code{"AQI"} color palette, 
as well as a suite of sequential and diverging palettes from the 
\pkg{RColorBrewer} R package.

The sequential palettes names are
 
\code{"Blues" "BuGn" "BuPu" "GnBu" "Greens" "Greys" "Oranges" "OrRd" "PuBu" 
 "PuBuGn" "PuRd" "Purples" "RdPu" "Reds" "YlGn" "YlGnBu" "YlOrBr" "YlOrRd"}

The diverging palettes are 

\code{"BrBG" "PiYG" "PRGn" "PuOr" "RdBu" "RdGy" "RdYlBu" "RdYlGn" "Spectral"}

Additional map tile info found at:
\url{http://maps.stamen.com/}
}
\examples{
\dontrun{
LA_basin <- pas_load() \%>\% pas_filterArea(-118.5, -117.5, 33.5, 34.5)
pas_staticMap(LA_basin, palette = "AQI", zoomAdjust = 1)
}

}
\keyword{pa_synoptic}
