% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_within.R
\name{pas_within}
\alias{pas_within}
\title{Find PurpleAir sensors within radial distance}
\usage{
pas_within(pas, label = NULL, radius = "1 km")
}
\arguments{
\item{pas}{PurpleAir \emph{pas} object.}

\item{label}{Target (centered) Purple Air sensor label.}

\item{radius}{Distance from target with unit (i.e "15 km").}
}
\value{
A subset of the given \emph{pas} object.
}
\description{
Filter for PurpleAir sensors within a specified distance from 
a target Purple Air sensor.
}
\details{
\code{radius} Should be a numeric string with provided metric unit separated by a space,
such as \code{"250 m"}.
}
\examples{
\dontrun{
pas_within(example_pas, "SaratogaMiddleHighSchool", "1024 km")
}
}
\seealso{
\link{pas_filter}

\link{pas_filterArea}
}
