% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_aggregate.R
\name{pat_aggregate}
\alias{pat_aggregate}
\title{Aggregation statistics for PurpleAir Timeseries objects}
\usage{
pat_aggregate(pat, FUN = mean, ...)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{FUN}{The function to be applied to each vector of numeric \code{pat} data.}

\item{...}{(Optional) Additional arguments for \code{FUN}.}
}
\value{
Returns a dataframe with aggregation statistics.
}
\description{
Calculates statistics associated with the aggregation of raw 
PurpleAir Timeseries data onto a regular time axis.

Temporal aggregation involves creating time period bins defined by
\code{period} and then calculating the statistics associated with the raw
data measurements that fall within each bin. The result is a dataframe with
a regular time axis and multiple columns of output for every column of
input.


FUN is found by a call to match.fun and typically is specified as a function or a symbol (e.g., a backquoted name) or a character string specifying a function to be searched for from the environment of the call to


Each of the \emph{pat} object data columns will result in the following
basic statistics:

\enumerate{
\item{\code{mean}}
\item{\code{median}}
\item{\code{sd}}
\item{\code{min}}
\item{\code{max}}
\item{\code{count}}
}

For the paired \code{pm25_A} and \code{pm25_B} data columns, the following 
additional statistics are generated by applying a two-sample t-test to the 
A and B channel data in each bin:

\enumerate{
\item{\code{pm25_t} -- t-test statistic}
\item{\code{pm25_p} -- p-value}
\item{\code{pm25_df} -- degrees of freedom}
}

These statistics are used to assign aggregate (hourly) values, (\emph{e.g.}
\code{temperature_mean}). They can also be used to invalidate some aggregate 
values based on \code{sd}, \code{count} or \code{pm25_t}.
}
\examples{
\dontrun{
avg_pat <- pat_aggregate(pat, mean, na.rm = TRUE)
}

}
\keyword{pa_timeseries}
