% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_calendarPlot.R
\name{pat_calendarPlot}
\alias{pat_calendarPlot}
\title{Calendar Heat Map Plot}
\usage{
pat_calendarPlot(pat = NULL, palette = NULL, ncol = 3,
  aspectRatio = 4/5, channel = "ab", qc_algorithm = "hourly_AB_01",
  min_count = 20)
}
\arguments{
\item{pat}{a pat object}

\item{palette}{a palette}

\item{ncol}{columns of plot}

\item{aspectRatio}{aspect ratio of the plot}

\item{channel}{Data channel to use for PM2.5 -- one of "a", "b or "ab".}

\item{qc_algorithm}{Named QC algorithm to apply to hourly aggregation stats.}

\item{min_count}{Aggregation bins with fewer than `min_count` measurements
will be marked as `NA`.}
}
\value{
ggobject
}
\description{
Plot a calendar heat map of daily PM2.5 average.

Current QC algorithms exist for \code{channel = "ab"} and include:
\itemize{
\item{\code{hourly_AB_00}}
\item{\code{hourly_AB_01}}
}
}
\note{
This function is currently optimized for annual time periods.
}
\examples{
\donttest{
pat_calendarPlot(pat = example_pat)
}

}
\seealso{
\link{PurpleAirQC_hourly_AB_00}

\link{PurpleAirQC_hourly_AB_01}
}
