% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_createASTimeseries.R
\name{pat_createASTimeseries}
\alias{pat_createASTimeseries}
\title{Create a AirSensor Timeseries object}
\usage{
pat_createASTimeseries(pat, period = "1 hour", ...)
}
\arguments{
\item{pat}{Purple Air timeseries data from \code{createPATimeseriesObject()}}

\item{period}{The time period to average to. Can be "sec", "min", "hour", 
"day", "DSTday", "week", "month", or "year". A number can also
precede these options followed by a space (i.e. "2 day" or "37 min").}

\item{...}{Extra parameters to fine tune \code{pat_aggregate()} function}
}
\value{
List with original \code{meta} and restructured \code{data} elements

"as_timeseries" list of aggregated time series PurpleAir data
}
\description{
Aggregates a PurpleAir timeseries object along the datetime
axis to produce a new dataframe with 13 columns of period-aggregated
timeseries data: averages, standard deviations, measurement frequency, and 
qualitity control.
}
\examples{
\dontrun{
pat <- 
  AirSensor::example_pat \%>\%
  pat_filterDate(20180701, 20180901)
ast <- pat_createASTimeseries(pat, "1 hour")
}
}
\seealso{
\link{pat_aggregate}
}
