% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_dailySoHIndexPlot.R
\name{pat_dailySoHIndexPlot}
\alias{pat_dailySoHIndexPlot}
\title{Daily State of Health metric plot}
\usage{
pat_dailySoHIndexPlot(pat = NULL)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}
}
\description{
This function plots as subset of the most useful State of Health 
metrics calculated by the \code{pat_dailySoH} function. The function 
runs \code{pat_dailySoH} internally and uses the output to create 
the plot.
}
\examples{

pat_dailySoHIndexPlot(example_pat_failure_A)


}
