% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_dailySoHPlot.R
\name{pat_dailySoHPlot}
\alias{pat_dailySoHPlot}
\title{Daily State of Health metric plot}
\usage{
pat_dailySoHPlot(pat = NULL)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}
}
\description{
This function plots as subset of the most useful State of Health 
metrics calculated by the \code{pat_dailyStateOfHealth} function. The function 
runs \code{pat_dailyStateOfHealth} internally and uses the output to create 
the plot.
}
