% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_dailyStateOfHealth.R
\name{pat_dailyStateOfHealth}
\alias{pat_dailyStateOfHealth}
\title{Daily state of health}
\usage{
pat_dailyStateOfHealth(pat = NULL, FUNs = c("SoH_dailyPctDC",
  "SoH_dailyPctReporting", "SoH_dailyPctValid", "SoH_dailyCorrelation"))
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{FUNs}{Vector of function names. All the passed in functions must output
tibbles with the same number of rows}
}
\description{
This function combines the output of the State of Health (SoH) 
function arguments into a single tibble. The output tibble should have
}
\examples{
 
SoH <- 
  example_pat_failure_B \%>\%
  pat_dailyStateOfHealth() 
head(SoH)
#timeseriesTbl_multiplot(tbl, ylim = c(0,100))
}
