% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paTimeseries.R
\name{pat_distinct}
\alias{pat_distinct}
\title{Retain only distinct data records in pat$data}
\usage{
pat_distinct(pat)
}
\arguments{
\item{pat}{\emph{pat} object}
}
\value{
A \emph{pat} object with no duplicated data records.
}
\description{
Performs two passes to guarantee that the \code{datetime} axis
contains no repeated values:

\enumerate{
\item{remove any duplicate records}
\item{guarantee that rows are in \code{datetime} order}
\item{average together fields for any remaining records that share the same
\code{datetime}}
}
}
