% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_downloadParseData.R
\name{pat_downloadParseData}
\alias{pat_downloadParseData}
\title{Download PurpleAir timeseries data}
\usage{
pat_downloadParseData(
  id = NULL,
  label = NULL,
  pas = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  baseURL = "https://api.thingspeak.com/channels/"
)
}
\arguments{
\item{id}{PurpleAir sensor 'deviceDeploymentID'.}

\item{label}{PurpleAir sensor 'label'.}

\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{startdate}{Desired start time (ISO 8601).}

\item{enddate}{Desired end time (ISO 8601).}

\item{timezone}{Timezone used to interpret start and end dates.}

\item{baseURL}{Base URL for Thingspeak API.}
}
\value{
List of type \code{pa_timeseries} containing \code{meta} and 
\code{data} elements with timeseries metadata and data, respectively.
}
\description{
Downloads timeseries data for a specific PurpleAir sensor 
from the ThingSpeak API and parses the content into a dataframe. This 
function will always return dataframe with the appropriate columns even if no 
data are returned from ThingSpeak.
}
\references{
https://www2.purpleair.com/community/faq
}
