% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_downloadParseRawData.R
\name{pat_downloadParseRawData}
\alias{pat_downloadParseRawData}
\title{Download time series data from PurpleAir}
\usage{
pat_downloadParseRawData(
  api_key = NULL,
  sensor_index = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = "UTC",
  average = 0,
  fields = SENSOR_HISTORY_PM25_FIELDS,
  baseUrl = "https://api.purpleair.com/v1/sensors"
)
}
\arguments{
\item{api_key}{PurpleAir API Read Key. If \code{api_key = NULL}, it
will be obtained using \code{getAPIKey("PurpleAir-read")}.
See \code{MazamaCoreUtils::\link[MazamaCoreUtils:setAPIKey]{setAPIKey}}.}

\item{sensor_index}{PurpleAir sensor unique identifier.}

\item{startdate}{Desired start datetime (ISO 8601).}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{timezone}{Olson timezone used to interpret dates.}

\item{average}{Temporal averaging in minutes performed by PurpleAir. One of:
0 (raw), 10, 30, 60 (hour), 360, 1440 (day).}

\item{fields}{Character string with PurpleAir field names for the Get Sensor Data API.}

\item{baseUrl}{Base URL for the PurpleAir API.}
}
\value{
Dataframe of time series PurpleAir data.
}
\description{
Download and parse time series data for a specific \code{sensor_index}.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

initializeMazamaSpatialUtils()

pat_raw <-
  pat_downloadParseRawData(
    api_key = MY_API_READ_KEY,
    sensor_index = "2323",
    startdate = "2023-02-01",
    enddate = "2023-02-02",
    timezone = "UTC",
    fields = SENSOR_HISTORY_PM25_FIELDS,
  )

View(pat_raw[1:100,])

}, silent = FALSE)
}
}
\references{
\href{https://www2.purpleair.com}{PurpleAir}

\href{https://api.purpleair.com}{PurpleAir API}

\href{https://www2.purpleair.com/policies/terms-of-service}{PurpleAir Terms of service}

\href{https://www2.purpleair.com/pages/license}{PurpleAir Data license}

\href{https://www2.purpleair.com/pages/attribution}{PurpleAir Data Attribution}
}
