% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_dygraph.R
\name{pat_dygraph}
\alias{pat_dygraph}
\title{Interactive time series plot}
\usage{
pat_dygraph(
  pat = NULL,
  parameter = "pm25",
  sampleSize = 5000,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  tlim = NULL,
  rollPeriod = 1,
  showLegend = TRUE,
  colors = NULL,
  timezone = NULL
)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object from \code{pat_createNew()}}

\item{parameter}{Data to display: "pm25", "humidity", "temperature" or "pressure".}

\item{sampleSize}{Either an integer or fraction to determine sample size.}

\item{title}{title text}

\item{xlab}{optional title for the x axis}

\item{ylab}{optional title for the y axis}

\item{tlim}{optional vector with start and end times (integer or character
representing YYYYMMDD[HH])}

\item{rollPeriod}{Width (hours) of rolling mean to be applied to the data.}

\item{showLegend}{Logical specifying whether to add a legend.}

\item{colors}{Vector of colors to be used for plotting.}

\item{timezone}{Olson timezone used to interpret \code{tlim}. (Defaults to 
\code{pat} local time.)}
}
\value{
Initiates the interactive dygraph plot in RStudio's 'Viewer' tab.
}
\description{
This function creates interactive graphs that will be displayed
in RStudio's 'Viewer' tab.

The list of available parameters include:

\itemize{
\item{\code{pm25} -- A and B channel PM2.5 (ug/m3)}
\item{\code{temperature} -- temperature (F)}
\item{\code{humidity} -- humidity (\%)}
\item{\code{pressure} -- pressure (hPa)}
}
}
\keyword{pa_timeseries}
