% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paTimeseries.R
\name{pat_extractDataFrame}
\alias{pat_extractDataFrame}
\alias{pat_extractData}
\alias{pat_extractMeta}
\title{Extract dataframes from \emph{pat} objects}
\usage{
pat_extractData(pat)

pat_extractMeta(pat)
}
\arguments{
\item{pat}{\emph{pat} object to extract dataframe from.}
}
\value{
A dataframe from the given \emph{pat} object
}
\description{
Extract dataframes from \emph{pat} objects
}
