% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_filterData.R
\name{pat_filterData}
\alias{pat_filterData}
\title{General purpose PurpleAir time series filter}
\usage{
pat_filterData(pat, ...)
}
\arguments{
\item{pat}{a pa_timeseries dataframe}

\item{...}{Logical predicates defined in terms of the variables in the 
`pat$data`. Multiple conditions are combined with & or seperated by a comma. 
Only rows where the condition evaluates to TRUE are kept.}
}
\value{
a pa_timeseries object
}
\description{
A generalized data filter for pa_timeseries (`pat`) objects to 
choose rows/cases where conditions are true. Rows where the condition 
evaluates to NA are dropped.
}
\examples{
\dontrun{
unhealthy <- pat_filterData(pat, pm25_A > 55.5, pm25_B > 55.5) 
}
}
\seealso{
\link{pat_filterDate}
}
