% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_findOutliers.R
\name{pat_findOutliers}
\alias{pat_findOutliers}
\title{Detect and Fix Timeseries Outliers}
\usage{
pat_findOutliers(pat, n = 11, thresholdMin = 6,
  replaceOutliers = FALSE, showPlot = TRUE)
}
\arguments{
\item{pat}{Purple Air Timeseries "pat" object from \code{createPATimeseriesObject()}}

\item{n}{integer window size}

\item{thresholdMin}{threshold value for outlier detection}

\item{replaceOutliers}{logical specifying whether replace outlier with the window median value}

\item{showPlot}{logical specifying whether to generate outlier detection plots}
}
\value{
\code{pat} timeseries object with outliers replaced by median values.
}
\description{
Outlier detection using a Median Average Deviation "Hampel" filter.

The \code{thresholdMin} level is similar to a sigma value for normally distributed data.
Hampel filter values above 6 indicate a data value that is extremely unlikely to
be part of a normal distribution (~ 1/500 million) and therefore very likely to
be an outlier. By choosing a relatively large value for thresholdMin we make it
less likely that we will generate false positives.

Specifying \code{replaceOutliers = TRUE} allows you to perform smoothing by 
replacing outliers with the window median value. Using this technique, you can
create an highly smoothed, artificial dataset by setting \code{thresholdMin = 1}
or lower (but always above zero).
}
\note{
Additional documentation on the algorithm is available in 
\code{seismicRoll::findOutliers()}.
}
