% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_internalData.R
\name{pat_internalData}
\alias{pat_internalData}
\title{Compare all variables in a Purple Air Timeseries object}
\usage{
pat_internalData(pat, use_adc0 = FALSE, use_rssi = TRUE,
  showPlot = TRUE)
}
\arguments{
\item{pat}{Purple Air Timeseries "pat" object from \code{createPATimeseriesObject()}}

\item{use_adc0}{logical specifying whether include the \code{adc0} parameter}

\item{use_rssi}{logical specifying whether include the \code{rssi} parameter}

\item{showPlot}{logical specifying whether to generate internal data plots}
}
\value{
Tibble portion of the \code{pat} object, subset.
}
\description{
Creates a multi-panel comparing all variables in the \code{pat} 
object. If any variables have no valid data, they are omitted from the plot.

A tibble is returned containing only the requested variables and omitting any
that have no valid data.
}
