% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paTimeseries.R
\name{pat_isEmpty}
\alias{pat_isEmpty}
\title{Test for an empty \emph{pat} object}
\usage{
pat_isEmpty(pat)
}
\arguments{
\item{pat}{\emph{pat} object}
}
\value{
\code{TRUE} if no data exist in \code{pat}, \code{FALSE} otherwise.
}
\description{
Convenience function for \code{nrow(pat$data) == 0}.
This makes for more readable code in functions that need to test for this.
}
\examples{
pat_isEmpty(example_pat)

}
