% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_load.R
\name{pat_load}
\alias{pat_load}
\title{Load Purple Air Timeseries data}
\usage{
pat_load(label = NULL, startdate = NULL, enddate = NULL,
  timezone = "America/Los_Angeles",
  baseUrl = "http://smoke.mazamascience.com/data/PurpleAir/pat")
}
\arguments{
\item{label}{Purple Air 'label'}

\item{startdate}{A starting date in ymd order.}

\item{enddate}{An ending date in ymd order}

\item{timezone}{Timezone used to interpret datestamp.}

\item{baseUrl}{Base URL for synoptic data.}
}
\value{
An object of class "pa_timeseries".
}
\description{
A pre-generated \code{pa_timeseries} object will be loaded for
the given time interval. Archived data for SCAQMD sensors go back to 
January, 2018.

The \code{startdate} and \code{enddate} can be in the following format:

\itemize{
\item{\code{"YYYYmmdd"}}
}

By default, the current month is loaded.
}
\examples{
\dontrun{
pat_load("SCNP_20", 20190411, 20190521)
  pat_multiplot()
}
}
\seealso{
\link{pat_loadMonth}
}
