% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_loadMonth.R
\name{pat_loadMonth}
\alias{pat_loadMonth}
\title{Load Purple Air Timeseries data}
\usage{
pat_loadMonth(name = NULL,
  datestamp = strftime(lubridate::today("America/Los_Angeles"),
  "\%Y\%m"), timezone = "America/Los_Angeles",
  baseUrl = "http://smoke.mazamascience.com/data/PurpleAir/pat")
}
\arguments{
\item{name}{Purple Air 'label'}

\item{datestamp}{A date string in ymd order.}

\item{timezone}{Timezone used to interpret datestamp.}

\item{baseUrl}{Base URL for synoptic data.}
}
\value{
An object of class "pa_timeseries".
}
\description{
A pre-generated \code{pa_timeseries} object will be loaded for
the given month. Archived data for SCAQMD sensors go back to January, 2018.

The \code{datestamp} can must be in the following format:

\itemize{
\item{\code{"YYYYmm"}}
}

By default, the current month is loaded.
}
\examples{
\dontrun{
pat_loadMonth("SCNP_20", 201905)
  pat_multiplot()
}
}
\seealso{
\link{pat_loadLatest}
}
