% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_multiplot.R
\name{pat_multiplot}
\alias{pat_multiplot}
\title{Display multiple plots on one page}
\usage{
pat_multiplot(pat = NULL, plottype = "all", plotlist = NULL,
  cols = NULL)
}
\arguments{
\item{pat}{Purple Air Timeseries "pat" object from \code{createPATimeseriesObject()}}

\item{plottype}{Quick-reference plot types: "all", "aux", "pm25"}

\item{plotlist}{a list() of any number of ggplot objects to plot on a single pane}

\item{cols}{Number of columns in the plot layout}
}
\description{
# A plotting function that uses ggplot2 to display multiple 
ggplot objects in a single pane. Can either be passed individual ggplot 
objects or a pat object and a plot type. 
Typical usage would be to supply \code{pat} and use the \code{plottype} 
argument to quickly display preformatted plots. 
\itemize{
\item{"all": pm25_A, pm25_B, temperature, humidity}
\item{"pm25": PM2.5 from Channel A and Channel B}
\item{"aux": auxillary data (temperature, humidity)}
}
}
\note{
Additional documentation of the multiplot algorithm is available at 
cookbook-r.com.
}
