% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_sample.R
\name{pat_sample}
\alias{pat_sample}
\title{Sample PurpleAir time series data}
\usage{
pat_sample(pat, sampleSize = NULL, sampleFraction = NULL,
  weight = NULL, setSeed = NULL)
}
\arguments{
\item{pat}{Purple Air Timeseries "pat" object}

\item{sampleSize}{a non-negative integer giving the number of rows to choose.}

\item{sampleFraction}{the fraction of rows to choose.}

\item{weight}{a vector of probability weights for obtaining the elements of the
vector being sampled.}

\item{setSeed}{an integer that sets random numbver generation. Can be used to 
reproduce sampling.}
}
\value{
\code{pat} timeseries object - subset.
}
\description{
A sampling function that accepts PurpleAir timeseries dataframes
and reduces them by randomly selecting distinct rows of the users choosen 
size.
}
\examples{
\dontrun{
pat <- pat_load()
subset <- pat_sample(pat, sampleSize=1000, setSeed=1234)
}

}
