% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_scatterplot.R
\name{pat_scatterplot}
\alias{pat_scatterplot}
\title{Compare all variables in a Purple Air Timeseries object}
\usage{
pat_scatterplot(pat, parameters = c("datetime", "pm25_A", "pm25_B",
  "temperature", "humidity"))
}
\arguments{
\item{pat}{Purple Air Timeseries "pat" object}

\item{parameters}{vector of parameters to include}
}
\value{
Tibble portion of the \code{pat} object, subset.
}
\description{
Compare all variables in a Purple Air Timeseries object
}
