% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_subdate.R
\name{pat_subdate}
\alias{pat_subdate}
\title{Subset PurpleAir time series intervals}
\usage{
pat_subdate(pat, sd, ed = NULL, days = NULL, weeks = NULL)
}
\arguments{
\item{pat}{Purple Air Timeseries "pat" object from \code{createPATimeseriesObject()}}

\item{sd}{A start-date following ISO 8601 date standard format}

\item{ed}{A end-date following ISO 8601 date standard format}

\item{days}{Number of days to include in the subdate interval}

\item{weeks}{Number of weeks to include in the subdate interval}
}
\value{
A subset "pat" list with \code{meta} and \code{data} elements
}
\description{
Subset PurpleAir time series intervals by date intervals, days, 
and weeks.
}
