% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_timeAverage.R
\name{pat_timeAverage}
\alias{pat_timeAverage}
\title{Time averages for PrupleAir time series}
\usage{
pat_timeAverage(pat = NULL, parameter = "pm25_A", period = "10 min",
  stats = "mean", dataThreshold = 0, fill = FALSE)
}
\arguments{
\item{pat}{PurpleAir Timeseries "pat" object from \code{pat_load()}}

\item{parameter}{The variable of timeseries data to aggergate, such as
"pm25_B", "temperature", etc.}

\item{period}{The time period to average to. Can be "sec", "min", "hour", 
"day", "DSTday", "week", "month", "quarter" or "year". A number can also
 precede these options followed by a space (i.e. "2 day" or "37 min").}

\item{stats}{The statistic to apply when aggregating the data; default is the 
mean. Can be one of "mean", "max", "min", "median", "frequency", 
"sd", "percentile".}

\item{dataThreshold}{A % of the data capture threshold. A value of 0 means 
that all data will be used in a particular period regardless of the number of
values avaliable. Conversely, a value of 100 means that all data will need to 
be present to proceed, else it is recorded as NA.}

\item{fill}{When time series are expanded data are ‘padded out’ with NA. To 
‘pad-out’ the additional data with the first row in each original time 
interval, choose fill = TRUE.}
}
\value{
Returns a dataframe with a date in class POSIXct.
}
\description{
Function to flexibly aggregate or expand data frames by 
different time periods and calculating vector-averages for a PurpleAir time 
series object. This function should be useful in many 
circumstances where it is necessary to work with different time average data.
}
\examples{
\dontrun{
# Hourly average of Channel A PM2.5 density measurement
pat_timeAverage(pat, "pm25_A", "1 hour")
}

\dontrun{
# Maximum weekly temperature
pat \%>\%
  pat_timeAverage("temperature", "1 week", stats = "max")
}
}
\keyword{pa_timeseries}
