% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAirQC_aggregationPlot.R
\name{PurpleAirQC_aggregationPlot}
\alias{PurpleAirQC_aggregationPlot}
\title{Faceted plot of aggregation statistics}
\usage{
PurpleAirQC_aggregationPlot(aggregationStats = NULL,
  parameterPattern = NULL, parameters = NULL, nrow = 5,
  autoRange = TRUE)
}
\arguments{
\item{aggregationStats}{PurpleAir Timeseries \emph{aggregationStats} object.}

\item{parameterPattern}{Pattern used to match groups of parameters.}

\item{parameters}{Custom vector of aggregation parameters to view.}

\item{nrow}{Number of rows in the faceted plot.}

\item{autoRange}{Logical specifying whether to scale the y axis separately
for each plot or to use a common y axis.}
}
\description{
A plotting function that uses ggplot2 to display a plot of each 
output category from the pat_aggregateOutlierCounts() function. Created to 
have a quick look at all the stats to help identify necessary quality control
methods for PurpleAir Timeseries \emph{aggregationStats} objects.
}
\examples{
\donttest{
aggregationStats <- pat_aggregateOutlierCounts(example_pat_failure_A)
PurpleAirQC_aggregationPlot(aggregationStats, 
                            parameterPattern = "humidity_m|temperature_m", 
                            nrow = 2)
}
}
