% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAirQC_validationPlot.R
\name{PurpleAirQC_validationPlot}
\alias{PurpleAirQC_validationPlot}
\title{Validation plot for QC algorithms}
\usage{
PurpleAirQC_validationPlot(pat = NULL, period = "1 hour",
  qc_algorithm = "hourly_AB_01", min_count = 20)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{period}{Time period to average to. Can be "sec", "min", "hour", 
"day", "DSTday", "week", "month", "quarter" or "year". A number can also
 precede these options followed by a space (i.e. "2 day" or "37 min").}

\item{qc_algorithm}{Named QC algorithm to apply to hourly aggregation stats.}

\item{min_count}{Aggregation bins with fewer than `min_count` measurements
will be marked as `NA`.}
}
\value{
A dataframe of aggregation statistics.
}
\description{
A diagnostic plot is generated showing time series for the
aggregation statistics used in the QC algorithm. Reviewing these time series
for different "pat" objects can provide insight into what kinds of instrument
problems a QC algorithm will catch or miss.

Current QC algorithms include:
\itemize{
\item{\code{hourly_AB_00}}
\item{\code{hourly_AB_01}}
}
}
\examples{
\dontrun{
scsg_15 <- pat_load("SCSG_15", "2019-06-13", "2019-06-20")
PurpleAirQC_validationPlot(scsg_15)

scap_14 <- pat_load("SCAP_14", "2019-06-13", "2019-06-20")
PurpleAirQC_validationPlot(scap_14)

scem_05 <- pat_load("SCEM_05", "2019-06-13", "2019-06-20")
PurpleAirQC_validationPlot(scem_05)
}
}
\seealso{
PurpleAirQC_hourly_AB_00

PurpleAirQC_hourly_AB_01
}
