% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAirSoH_dailyABFit_OLD.R
\name{PurpleAirSoH_dailyABFit_OLD}
\alias{PurpleAirSoH_dailyABFit_OLD}
\title{Daily linear model fit values}
\usage{
PurpleAirSoH_dailyABFit_OLD(pat = NULL, aggregationPeriod = "10 min")
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{aggregationPeriod}{The period of time over which to aggregate the data. 
in order to create complete pairs between the A and B channels. If there is 
not a complete pair, the function will return an "NA" for that day. It is
recommended that this value is between "2 min" and "1 hour" The smaller the 
\emph{aggregation period}, the longer this function will take to run. The 
default is set to "10 min" for this reason.}
}
\description{
This function calculates daily linear model values between
the \code{pm25_A} and \code{pm25_B} channels. A daily r-squared value is
returned in addition to the coefficients of the linear fit (slope and 
intercept)
}
\examples{
 
tbl <- 
  example_pat_failure_A \%>\%
  PurpleAirSoH_dailyABFit_OLD(aggregationPeriod = "30 min") 
  
timeseriesTbl_multiplot(
  tbl, 
  ylim = c(-1,1), 
  style = "line"
)

}
