% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAirSoH_dailyCorrelation.R
\name{PurpleAirSoH_dailyCorrelation}
\alias{PurpleAirSoH_dailyCorrelation}
\title{Daily correlation values}
\usage{
PurpleAirSoH_dailyCorrelation(pat = NULL, aggregationPeriod = "10 min")
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{aggregationPeriod}{The period of time over which to aggregate the data. 
in order to create complete pairs between the A and B channels. If there is 
not a complete pair, the function will return an "NA" for that day. It is
recommended that this value is between "2 min" and "1 hour" The smaller the 
\emph{aggregation period}, the longer this function will take to run. The 
default is set to "10 min" for this reason.}
}
\description{
This function calculates the daily correlation values between
the \code{pm25_A}, \code{pm25_B}, \code{humidity}, and \code{temperature} 
channels. One correlation value for each channel pair will 
be returned for each day. A linear fit between channels A and B is also 
calculated for each day and the coefficients are returned as the slope and 
the intercept.
}
\examples{
 
tbl <- 
  example_pat_failure_A \%>\%
  PurpleAirSoH_dailyCorrelation(aggregationPeriod = "30 min") 
  
timeseriesTbl_multiplot(
  tbl, 
  parameters = c("pm25_A_pm25_B_cor", "temperature_humidity_cor", "pm25_A_temperature_cor"),
  ylim = c(-1,1), 
  style = "line"
)

timeseriesTbl_multiplot(
  tbl, 
  parameterPattern = "_A.*_B",
  autoRange = TRUE, 
  style = "line"
)

}
