% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAirSoH_dailyCorrelation_OLD.R
\name{PurpleAirSoH_dailyCorrelation_OLD}
\alias{PurpleAirSoH_dailyCorrelation_OLD}
\title{Daily correlation values}
\usage{
PurpleAirSoH_dailyCorrelation_OLD(pat = NULL,
  aggregationPeriod = "10 min")
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{aggregationPeriod}{The period of time over which to aggregate the data. 
in order to create complete pairs between the A and B channels. If there is 
not a complete pair, the function will return an "NA" for that day. It is
recommended that this value is between "2 min" and "1 hour" The smaller the 
\emph{aggregation period}, the longer this function will take to run. The 
default is set to "10 min" for this reason.}
}
\description{
This function calculates the daily correlation values between
the \code{pm25_A}, \code{pm25_B}, \code{humidity}, and \code{temperature} 
channels. One correlation value for each channel pair except \code{pm25_A}, 
\code{pm25_B}, and \code{humidity}, \code{temperature} will be returned for 
each day. All returned values are expected to hover near 0 for a properly
functioning sensor.
}
\examples{
 
tbl <- 
  example_pat_failure_A \%>\%
  PurpleAirSoH_dailyCorrelation_OLD(aggregationPeriod = "30 min") 
  
timeseriesTbl_multiplot(
  tbl, 
  parameters = c("pm25_B_temperature_cor", "pm25_A_temperature_cor"),
  ylim = c(-1,1), 
  style = "line"
)

timeseriesTbl_multiplot(
  tbl, 
  autoRange = TRUE, 
  style = "line"
)

}
