% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAirSoH_dailyOtherFit.R
\name{PurpleAirSoH_dailyOtherFit}
\alias{PurpleAirSoH_dailyOtherFit}
\title{Daily fit values}
\usage{
PurpleAirSoH_dailyOtherFit(pat = NULL)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}
}
\description{
This function calculates the daily linear model values between
the \code{pm25_A}, \code{pm25_B}, \code{humidity}, and \code{temperature} 
channels. One r-squared value for each channel pair except \code{pm25_A}, 
\code{pm25_B}, and \code{humidity}, \code{temperature} will be returned for 
each day. All returned values are expected to hover near 0 for a properly
functioning sensor.
}
\examples{
 
tbl <- 
  example_pat_failure_A \%>\%
  PurpleAirSoH_dailyOtherFit() 
  
timeseriesTbl_multiplot(
  tbl, 
  parameters = c("pm25_B_temperature_rsquare", "pm25_A_temperature_rsquare"),
  ylim = c(-1,1), 
  style = "line"
)

timeseriesTbl_multiplot(
  tbl, 
  autoRange = TRUE, 
  style = "line"
)

}
