% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAirSoH_dailyToIndex_00.R
\name{PurpleAirSoH_dailyToIndex_00}
\alias{PurpleAirSoH_dailyToIndex_00}
\title{Daily state of health index}
\usage{
PurpleAirSoH_dailyToIndex_00(
  SoH = NULL,
  minPctReporting = 50,
  breaks = c(0, 0.2, 0.8, 1)
)
}
\arguments{
\item{SoH}{PurpleAir daily State-of-Health dataframe.}

\item{minPctReporting}{Percent reporting threshold for A and B channels.}

\item{breaks}{Breaks used to convert index values into index bins.}
}
\description{
This function calculates a multi-metric index based on the data
in \code{SoH} dataframe passed in. A tibble is returned containing a state of 
health index for each day. The returned tibble contains columns: 
\code{datetime}, \code{index}, and \code{index_bin}.

The \code{index} column contains a value normalized between 
0 and 1 where 0 represents low confidence in the sensor data and 1 represents 
high confidence. The \code{index_bin} is one of 1, 2, or 3 and represents
poor, fair, and good data respectively. 

The \code{index} is calculated in the following manner:

\enumerate{
\item{If the A or B channel percent reporting is < \code{minPctReporting}, \code{index = 0}}
\item{Otherwise, \code{index = pm25_A_pm25_B_rsquared}}
}

The \code{breaks} are used to convert \code{index} into the \code{indenx_bin}
poor-fair-good values.
}
\examples{
 
library(AirSensor)

tbl <- 
  example_pat_failure_A \%>\%
  pat_dailySoH() \%>\%
  PurpleAirSoH_dailyToIndex_00() 
  
head(tbl)
}
