% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwfsl_load.R
\name{pwfsl_load}
\alias{pwfsl_load}
\title{Get PWFSLSmoke Monitoring Data}
\usage{
pwfsl_load()
}
\value{
List with \code{meta} and \code{data} elements
}
\description{
Loads recent PM2.5 monitoring data from the US Forest Service
Pacific Wildland Fire Sciences Lab. This function performs the same data loading 
step as \code{pwfsl_loadLatest}, but has a shorter name for consistency within 
the \code{MazamaPurpleAir} package. By default, this loads data from all 50 states
for the past 10 days. 

This function is a wrapper around \code{PWFSLSmoke::monitor_loadLatest}.

Data for the most recent 45 days can be downloaded using \code{monitor_load}
in the PWFSLSmoke package, whose value will be a /code{ws_monitor} object with 
the same structuer as the value returned by \code{pwfsl_loadLatest}.
}
\examples{
\dontrun{
pwfsl <- pwfsl_load()
}
}
