% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\name{removeArchiveBaseDir}
\alias{removeArchiveBaseDir}
\title{Remove data archive base directory}
\usage{
removeArchiveBaseDir()
}
\value{
Silently returns previous value of the base directory.
}
\description{
Resets the data archive base directory to NULL. Used for internal 
testing.
}
\seealso{
ArchiveBaseDir

getArchiveBaseDir

setArchiveBaseDir
}
\keyword{environment}
\keyword{internal}
