% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\name{removeArchiveBaseUrl}
\alias{removeArchiveBaseUrl}
\title{Remove data archive base URL}
\usage{
removeArchiveBaseUrl()
}
\value{
Silently returns previous value of the base URL.
}
\description{
Resets the data archive base URL to NULL. Used for internal 
testing.
}
\seealso{
ArchiveBaseUrl

getArchiveBaseUrl

setArchiveBaseUrl
}
\keyword{environment}
\keyword{internal}
