% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gen.R
\name{replaceOutliers}
\alias{replaceOutliers}
\title{Replace outliers with rolling median}
\usage{
replaceOutliers(df, parameter = NULL, medWin = 7, ...)
}
\arguments{
\item{df}{A data frame.}

\item{parameter}{The data frame parameter to check and replace outliers}

\item{medWin}{The size of the rolling median window}

\item{...}{Parameters to extend \code{flagOutliers}}
}
\value{
a \code{data.frame} with replaced outliers
}
\description{
Perform smoothing by replacing outliers with the window median 
value. Using this technique, you can create an highly smoothed, artificial 
dataset by setting \code{thresholdMin = 1} or lower (but always above zero).
}
