% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-SCAQMD.R
\docType{data}
\name{SCAQMD_SENSOR_INDICES}
\alias{SCAQMD_SENSOR_INDICES}
\title{Comma separated list of SCAQMD sensor indices.}
\format{
String with comma-separated sensor_index values.
}
\usage{
SCAQMD_SENSOR_INDICES
}
\description{
Character string with PurpleAir \code{sensor_index} values
maintained by the South Coast Air Quality Management District (SCAQMD).
These can be used with \code{pas_createNew(show_only = ...)} to create a 
\emph{pas} object containing only SCAQMD sensors.
}
\keyword{datasets}
