% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{scatterplot}
\alias{scatterplot}
\title{Matrix scatter plot variables in a data frame}
\usage{
scatterplot(data, parameters = NULL, sampleSize = 5000,
  sampleFraction = NULL, shape = 18, size = 1.5, color = "black",
  alpha = 0.5)
}
\arguments{
\item{data}{data frame}

\item{shape}{symbol to use for points}

\item{size}{size of points}

\item{color}{color of points}

\item{alpha}{opacity of points}
}
\description{
Creates a multi-panel scatterplot comparing all variables in the
data frame object. If any variables have not valid data, they are omitted 
from the plot.
}
