% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_aggregate.R
\name{sensor_aggregate}
\alias{sensor_aggregate}
\title{Aggregate airsensor Object}
\usage{
sensor_aggregate(
  sensor,
  FUN = function(x) {     mean(x, na.rm = TRUE) },
  unit = "hours",
  count = 24
)
}
\arguments{
\item{sensor}{Sensor/ws_monitor \emph{airsensor} object.}

\item{FUN}{The function to be applied to each vector of numeric \code{sensor} data.}

\item{unit}{Character string specifying temporal units for binning.}

\item{count}{Number of units per bin.}
}
\value{
Returns an aggregated \emph{airsensor} object.
}
\description{
Aggregate (\emph{airsensor}) object along its
datetime axis. Temporal aggregation involves splitting a \emph{airsensor} object into
separate bins along its datetime axis. \code{FUN} is mapped to the \emph{airsensor}
numeric variables in each bin, which are then recombined into an aggregated
\emph{sensor} object containing the same metadata as the incoming \code{airsensor}.
The functionality extends that of \code{pat_aggregate()}.
}
\details{
\code{FUN} must operate on univariate numeric vectors and return a
scalar value. Besides the data variable, no additional arguments will be
provided to this function. This means that functions like \code{mean} and
\code{max} will need to be wrapped in a function that specifies
\code{na.rm = TRUE}. See the examples below.
}
\examples{
library(AirSensor)

sensor_aggregate(example_sensor)
}
\keyword{airsensor}
