% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_calendarPlot.R
\name{sensor_calendarPlot}
\alias{sensor_calendarPlot}
\title{Calendar Heat Map Plot}
\usage{
sensor_calendarPlot(sensor = NULL, palette = NULL, ncol = 3,
  aspectRatio = 4/5, discrete = TRUE)
}
\arguments{
\item{sensor}{an 'airsensor' object}

\item{palette}{a palette}

\item{ncol}{columns of plot}

\item{aspectRatio}{aspect ratio of the plot}

\item{discrete}{a boolean to determine if color breaks will be discrete by a 
predefined scale, or continous based on the data domain.}
}
\value{
ggobject
}
\description{
Plot a calendar heat map of daily PM2.5 average.
}
\note{
This function is currently optimized for annual time periods.
}
\examples{
\donttest{
sensor_calendarPlot(sensor = example_sensor)
}

}
\seealso{
\link{pat_calendarPlot}
}
