% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\docType{data}
\name{SENSOR_DATA_AVG_PM25_FIELDS}
\alias{SENSOR_DATA_AVG_PM25_FIELDS}
\title{Comma-separated list of fields needed to create a \emph{pas} object.}
\format{
String with comma-separated field names
}
\usage{
SENSOR_DATA_AVG_PM25_FIELDS
}
\description{
Character string with PurpleAir field names used in
\code{pas_downloadParseRawData()}. These fields include most of the
"information and status" fields, "humidity", "temperature", "pressure" and
simple running average PM2.5 fields for different time periods.
}
\references{
\href{https://api.purpleair.com/#api-sensors-get-sensor-data}{Get Sensor Data API}
}
\keyword{datasets}
