% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_filter.R
\name{sensor_filter}
\alias{sensor_filter}
\title{General purpose data filtering for AirSensor objects}
\usage{
sensor_filter(sensor = NULL, ...)
}
\arguments{
\item{sensor}{An AirSensor object.}

\item{...}{Logical predicates defined in terms of the variables in the 
\emph{sensor}.}
}
\value{
A subset of the incoming \emph{sensor}.
}
\description{
A generalized data filter for \emph{sensor} objects to 
choose rows/cases where conditions are true.  Multiple conditions are 
combined with \code{&} or seperated by a comma. Only rows where the condition 
evaluates to TRUE are kept.Rows where the condition evaluates to \code{NA}
are dropped.
}
\examples{
\dontrun{
monitorID <- example_sensor$meta$monitorId
twenties <- sensor_filter(example_sensor, SCAN_14 >= 20, SCAN_14 < 30)
head(twenties$data)
}

}
\seealso{
\link{sensor_filterDate}
}
