% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_API.R
\docType{data}
\name{SENSOR_HISTORY_PM25_FIELDS}
\alias{SENSOR_HISTORY_PM25_FIELDS}
\title{Comma-separated list of fields needed for PM2.5 data analysis.}
\format{
String with comma-separated field names
}
\usage{
SENSOR_HISTORY_PM25_FIELDS
}
\description{
Character string with default PurpleAir field names used in
\code{pat_downloadParaseRawData()}. These fields are sufficient for most
QC algorithms and include most of the "information and status" fields,
"humidity", "temperature", "pressure" and the PM2.5 "pseudo average" fields.
}
\references{
\href{https://api.purpleair.com/#api-sensors-get-sensor-history-csv}{Get Sensor History API}
}
\keyword{datasets}
