% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sensor.R
\name{sensor_isEmpty}
\alias{sensor_isEmpty}
\title{Test for an empty \emph{sensor} object}
\usage{
sensor_isEmpty(sensor)
}
\arguments{
\item{sensor}{\emph{sensor} object}
}
\value{
\code{TRUE} if no data exist in \code{sensor}, \code{FALSE} otherwise.
}
\description{
Convenience function for \code{nrow(sensor$meta) == 0}.
This makes for more readable code in functions that need to test for this.
}
\examples{
example_sensor <- pat_createAirSensor(example_pat)
sensor_isEmpty(example_sensor)

}
