% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_videoFrame.R
\name{sensor_videoFrame}
\alias{sensor_videoFrame}
\title{Generate video frames}
\usage{
sensor_videoFrame(sensor = sensor_load(), communityRegion = NULL,
  frameTime = NULL, timeInfo = NULL, map = NULL)
}
\arguments{
\item{sensor}{An AirSensor object.}

\item{communityRegion}{The area to display.}

\item{frameTime}{POSIXct specifying the time to use.}

\item{timeInfo}{???.}

\item{map}{A pre-generated basemap image.}
}
\value{
ret
}
\description{
desc
}
