% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\name{setArchiveBaseUrl}
\alias{setArchiveBaseUrl}
\title{Set data archive base URL}
\usage{
setArchiveBaseUrl(archiveBaseUrl)
}
\arguments{
\item{archiveBaseUrl}{Base URL pointing to an archive of pre-generated data files.}
}
\value{
Silently returns previous value of base URL.
}
\description{
Sets the package base URL pointing to an archive of
pre-generated data files.
}
\seealso{
ArchiveBaseUrl

getArchiveBaseUrl
}
\keyword{environment}
