% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoH_dailyCorrelation.R
\name{SoH_dailyCorrelation}
\alias{SoH_dailyCorrelation}
\title{Daily correlation values}
\usage{
SoH_dailyCorrelation(pat = NULL, aggregationPeriod = "10 min")
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{aggregationPeriod}{The period of time over which to aggregate the data. 
in order to create complete pairs between the A and B channels. If there is 
not a complete pair, the function will return an "NA" for that day. It is
recommended that this value is between "2 min" and "1 hour" The smaller the 
\emph{aggregation period}, the longer this function will take to run. The 
default is set to "10 min" for this reason.}
}
\description{
This function calculates the daily correlation values between
the \code{pm25_A}, \code{pm25_B}, \code{humidity}, and \code{temperature} 
channels. One correlation value for each channel pair will 
be returned for each day.
}
\examples{
 
tbl <- 
  example_pat \%>\%
  SoH_dailyCorrelation(aggregationPeriod = "30 min") 
  
timeseriesTbl_multiplot(tbl, ylim = c(-1,1), style = "line")

}
