% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoH_dailyPctValid.R
\name{SoH_dailyPctValid}
\alias{SoH_dailyPctValid}
\title{Daily valid percentage}
\usage{
SoH_dailyPctValid(pat = NULL)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}
}
\description{
The number of valid (\emph{i.e.}, not NA or out-of-spec) sensor 
measurements are summed over the course of a calendar day (24 
hours unless a partial day is included in the data), then divided by the 
total number of measurements the sensor actually recorded in during that day 
(including NA and out-of-spec values) to return a percentage of the total
recorded measurements that are considered plausible.
}
\examples{
 
tbl <- 
  example_pat_failure_B \%>\%
  SoH_dailyPctValid() 

timeseriesTbl_multiplot(tbl, ylim = c(0,100))
}
