% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoH_pctDC.R
\name{SoH_pctDC}
\alias{SoH_pctDC}
\title{Daily DC Signal percentage}
\usage{
SoH_pctDC(pat = NULL, aggregation_period = "30 min", timezone = NULL)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{aggregation_period}{The period to aggregate over. Choose from "30 min",
"1 hour", or "1 day".}

\item{timezone}{Olson timezone at the location of interest}
}
\description{
This function calculates the daily percentage of DC signal 
recorded by the \code{pm25_A}, \code{pm25_B}, \code{humidity}, and 
\code{temperature} channels. The data are flagged as DC signal when the 
standard deviation of the data from each channel equals zero over the 
\code{aggregation_period}. The number of DC hours are summed over the day and
a daily DC percentage is returned.
}
\examples{
 
tbl <- 
  example_pat_failure_B \%>\%
  SoH_pctDC(aggregation_period = "30 min", timezone = "America/Los_Angeles") 

}
