% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoH_metrics.R
\name{SoH_pctReporting}
\alias{SoH_pctReporting}
\title{Daily reporting percentage}
\usage{
SoH_pctReporting(pat = NULL, samplingInterval = 120)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{samplingInterval}{The number of seconds between samples when the sensor 
is operating optimally.}
}
\description{
The number of sensor readings recorded per hour are summed over 
the course of a calendar day (24 hours unless the \code{pat} datetime column
includes a partial day at the beginning or end). This ishen divided by the 
number of samples the sensor would record in an ideal day 
(\code{24 * 3600 / samplingInterval}) to return a percentage of each 
day that the sensor is reporting data.
}
\note{
Purple Air II sensors reporting after the June, 2019 firmware
upgrade report data every 120 seconds. Prior to the upgrade, data were 
reported every 80 seconds.
}
\examples{
tbl <- 
  example_pat \%>\%
  SoH_pctReporting(80) 

timeseriesTbl_multiplot(tbl)
}
