% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoH_pctValid.R
\name{SoH_pctValid}
\alias{SoH_pctValid}
\title{Daily valid percentage}
\usage{
SoH_pctValid(pat = NULL)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}
}
\description{
The number of valid (i.e., not NA or out of spec) sensor 
readings recorded per hour are summed over the course of a calendar day (24 
hours unless a partial day is included in the data), then divided by the 
total number of samples the sensor actually recorded in during that day 
(including NA and out of spec values) to return a percentage of each 
day that the sensor recorded valid measurements.
}
\examples{
 
tbl <- 
  example_pat_failure_B \%>\%
  SoH_pctValid() 

}
