% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeMazamaSpatialUtils.R
\name{spatialIsInitialized}
\alias{spatialIsInitialized}
\title{Check if MazamaSpatialUtils has been initialized}
\usage{
spatialIsInitialized()
}
\value{
Logical.
}
\description{
Logical convenience function to check if 
\code{initializeMazamaSpatialUtils()} has been run.
}
\keyword{internal}
