% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_load.R
\name{wind_load}
\alias{wind_load}
\title{Load wind data}
\usage{
wind_load(monitorID = NULL, startdate = NULL, enddate = NULL,
  days = 7, timezone = "America/Los_Angeles",
  baseUrl = "http://smoke.mazamascience.com/data/PurpleAir/wind")
}
\arguments{
\item{monitorID}{A monitor ID}

\item{startdate}{Desired start datetime (ISO 8601)}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{days}{Number of days of data to include.}

\item{timezone}{Timezone used to interpret datestamp.}

\item{baseUrl}{Base URL for synoptic data.}
}
\value{
a data frame
}
\description{
Load pre-generated wind data for the given interval. Archived 
data for SCAQMD sensors go back to January, 2018.
}
\examples{
\dontrun{
wind_load(
  monitorID = "060950004_01",
  startdate = "20180521", 
  enddate = "20180627"
) 
}

}
