% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_loadMonth.R
\name{wind_loadMonth}
\alias{wind_loadMonth}
\title{Load monthly wind data}
\usage{
wind_loadMonth(monitorID = NULL, datestamp = NULL,
  timezone = "America/Los_Angeles",
  baseUrl = "http://smoke.mazamascience.com/data/PurpleAir/wind")
}
\arguments{
\item{monitorID}{A monitorID.}

\item{datestamp}{Date string in ym or ymd order}

\item{timezone}{Timezone used to interpret datestamp.}

\item{baseUrl}{Base URL for wind data.}
}
\value{
A data frame with columns \code{"date", "wd", "ws"}.
}
\description{
Load pre-generated wind data for the given month. Archived data 
appropriate for use with SCAQMD sensors go back to January, 2018.
}
\examples{
\dontrun{
wind_loadMonth(monitorID = "060950004_01", datestamp = "201805")
}

}
