% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\docType{data}
\name{ArchiveBaseUrl}
\alias{ArchiveBaseUrl}
\title{Base URL for pre-generated data}
\format{URL string.}
\description{
This package maintains an internal archive base URL which users 
must set using \code{setArchiveBaseUrl()}.

Known base URLs include:
\itemize{
\item{http://smoke.mazamascience.com/data/PurpleAir}
}

Package functions that load pre-generated data files download data from this
URL. These functions include:

\itemize{
\item{\code{pas_load()}}
\item{\code{pat_load()}}
\item{\code{pat_loadLatest()}}
\item{\code{pat_loadMonth()}}
\item{\code{sensor_load()}}
\item{\code{sensor_loadLatest()}}
\item{\code{sensor_loadMonth()}}
}
}
\seealso{
getArchiveBaseUrl

setArchiveBaseUrl
}
\keyword{environment}
