% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPATimeseriesObject.R
\name{createPATimeseriesObject}
\alias{createPATimeseriesObject}
\title{Create a Purple Air Timeseries object}
\usage{
createPATimeseriesObject(pat_raw = NULL)
}
\arguments{
\item{pat_raw}{raw Purple Air timeseries data from \code{downloadParseTimeseriesData()}}
}
\value{
List with original \code{meta} and restructured \code{data} elements

"pa_timeseries" list of time series PurpleAir data
}
\description{
The 'data' dataframe is converted from 'long format' with temporal
resolution of seconds to 'wide format' with temporal resolution in minutes.
In the process, the following columns of data are omitted:
\itemize{
\item{\code{pm1_atm}}
\item{\code{pm2.5_atm}}
\item{\code{pm10_atm}}
}
}
\examples{
\dontrun{
initializeMazamaSpatialUtils()
pas <- pas_load()
pat_raw <- downloadParseTimeseriesData(pas, name ='North Bend Weather', startdate = 20181908)
pat <- createPATimeseriesObject(pat_raw)
}
}
\seealso{
\link{downloadParseTimeseriesData}
}
