% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadParseTimeseriesData.R
\name{downloadParseTimeseriesData}
\alias{downloadParseTimeseriesData}
\title{Download Purple Air timeseries data}
\usage{
downloadParseTimeseriesData(pas = NULL, label = NULL, id = NULL,
  startdate = NULL, enddate = NULL,
  baseURL = "https://api.thingspeak.com/channels/")
}
\arguments{
\item{pas}{Purple Air 'enhanced' synoptic data}

\item{label}{Purple Air 'label'}

\item{id}{Purple Air 'ID'}

\item{startdate}{Desired UTC start time (ISO 8601)}

\item{enddate}{Desired UTC end time (ISO 8601)}

\item{baseURL}{Base URL for Thingspeak API}
}
\value{
The monitor time series given broken up into metadata and readings data
}
\description{
Timeseries data from a specific PurpleAir can be retrieved from the Thingspeak API .
}
