% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\docType{data}
\name{example_pas}
\alias{example_pas}
\title{Example Purple Air Synoptic dataset}
\format{A tibble with 7113 rows and 35 columns of data.}
\source{
https://www.purpleair.com/json
}
\description{
The \code{example_pas} dataset provides a quickly loadable
version of a \emph{pa_synoptic} object for practicing and code development.
This dataset was generated by running \code{pas_load()} on 2019-03-14.
}
\seealso{
example_raw_pas
}
\keyword{datasets}
