% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_pas}
\alias{example_pas}
\title{Example Purple Air Synoptic dataset}
\format{
A tibble with 16584 rows and 44 columns of data.
}
\source{
https://www.purpleair.com/json?all=true
}
\usage{
example_pas
}
\description{
The \code{example_pas} dataset provides a quickly loadable
version of a \emph{pa_synoptic} object for practicing and code examples.
This dataset was generated on 2020-09-15 by running:

\preformatted{
library(AirSensor)

initializeMazamaSpatialUtils()

example_pas <- pas_createNew(countryCodes = "US")

save(example_pas, file = "data/example_pas.rda")
}
}
\seealso{
example_pas_raw
}
\keyword{datasets}
