% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pas}
\alias{example_pas}
\title{Example Purple Air Synoptic dataset}
\format{
A tibble with 10220 rows and 45 columns of data.
}
\source{
https://www.purpleair.com/json
}
\usage{
example_pas
}
\description{
The \code{example_pas} dataset provides a quickly loadable
version of a \emph{pa_synoptic} object for practicing and code examples.
This dataset was generated on 2020-01-03 by running:

\preformatted{
library(AirSensor)

initializeMazamaSpatialUtils()

example_pas <- pas_createNew()

save(example_pas, file = "data/example_pas.rda")
}
}
\seealso{
example_pas_raw
}
\keyword{datasets}
