% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\docType{data}
\name{example_pat}
\alias{example_pat}
\title{Example Purple Air Timeseries dataset}
\format{An S3 object composed of "meta" and "data" data.}
\source{
https://www.purpleair.com/json
}
\description{
The sample_pas dataset provides a quickly loadable version of
a \emph{pa_synoptic} object for practicing and code development. The dataset 
was generated by running \code{pat_load()} on 2019-03-22 focused on a monitor
in central Seattle over a two-month interval, 2018-07-01 to 2018-09-01.
}
\seealso{
createPATimeseriesObject
}
\keyword{datasets}
