% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_pat}
\alias{example_pat}
\title{Example PurpleAir Timeseries dataset}
\format{
An S3 object composed of "meta" and "data" data.
}
\usage{
example_pat
}
\description{
The \code{example_pat} dataset provides a quickly loadable version of
a \emph{pa_timeseries} object for practicing and code examples.
This dataset was was generated on 2020-09-15 by running:

\preformatted{
library(AirSensor)

initializeMazamaSpatialUtils()

example_pat <- pat_createNew(
  id = "ebcb53584e44bb6f_3218",
  pas = example_pas,
  startdate = "2018-08-01",
  enddate = "2018-08-28",
  verbose = TRUE
)

save(example_pat, file = "data/example_pat.rda")
}
}
\seealso{
example_pat_failure_A

example_pat_failure_B
}
\keyword{datasets}
