% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_pat}
\alias{example_pat}
\title{Example PurpleAir Timeseries dataset}
\format{
An S3 object composed of "meta" and "data" data.
}
\usage{
example_pat
}
\description{
The \code{example_pat} dataset provides a quickly loadable version of
a \emph{pa_timeseries} object for practicing and code examples.
This dataset was was generated on 2023-03-16 by running:

\preformatted{
library(AirSensor)

initializeMazamaSpatialUtils()

example_pat <-
  pat_createNew(
    api_key = MY_API_READ_KEY,
    pas = example_pas,
    sensor_index = "3515",
    startdate = "2022-07-01",
    enddate = "2022-07-08",
    timezone = "UTC",
    verbose = TRUE
  )

save(example_pat, file = "data/example_pat.rda")
}
}
\keyword{datasets}
