% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pat}
\alias{example_pat}
\title{Example Purple Air Timeseries dataset}
\format{An S3 object composed of "meta" and "data" data.}
\usage{
example_pat
}
\description{
The \code{example_pat} dataset provides a quickly loadable version of
a \emph{pa_timeseries} object for practicing and code examples.
This dataset was was generated on 2019-07-06 by running:

\preformatted{
  pas <- pas_load()
  example_pat <- pat_createNew(pas, label = "Seattle",
                               startdate = "2018-07-01",
                               enddate = "2018-09-01")
}
}
\seealso{
example_pat_failure_A

example_pat_failure_B
}
\keyword{datasets}
