% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\docType{data}
\name{example_pat}
\alias{example_pat}
\title{Example Purple Air Timeseries dataset}
\format{An S3 object composed of "meta" and "data" data.}
\description{
The example_pat dataset provides a quickly loadable version of
a \emph{pa_timeseries} object for practicing and code development. This dataset 
was generated by running \code{pat_load()} for on a monitor
in central Seattle over a two-month interval, 2018-07-01 to 2018-09-01.
}
\seealso{
createPATimeseriesObject
}
\keyword{datasets}
