% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\docType{data}
\name{example_pat_failure_A}
\alias{example_pat_failure_A}
\title{Example Purple Air Timeseries dataset exhibiting moderate errors}
\format{An S3 object composed of "meta" and "data" data.}
\description{
The example_pat_failure_A dataset provides a quickly loadable
version of a \emph{pa_timeseries} object for practicing and code examples. 
This dataset was was generated on 2019-07-06 by running:

\preformatted{
  pas <- pas_load()
  example_pat_failure_A <- 
    pat_loadLatest(pas, name = "SCNP_20",
                   startdate = "2019-04-01",
                   enddate = "2019-04-18")
}
}
\seealso{
example_pat

example_pat_failure_B
}
\keyword{datasets}
