% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pat_failure_A}
\alias{example_pat_failure_A}
\title{Example PurpleAir Timeseries dataset exhibiting moderate errors}
\format{
An S3 object composed of "meta" and "data" data.
}
\usage{
example_pat_failure_A
}
\description{
The \code{example_pat_failure_A} dataset provides a quickly loadable
version of a \emph{pa_timeseries} object for practicing and code examples.
This dataset was was generated on 2020-09-15 by running:

\preformatted{
library(AirSensor)

initializeMazamaSpatialUtils()

example_pat_failure_A <- pat_createNew(
  label = "SCNP_20",
  pas = example_pas,
  startdate = "2019-04-01",
  enddate = "2019-04-18",
  verbose = "TRUE"
)

save(example_pat_failure_A, file = "data/example_pat_failure_A.rda")
}
}
\seealso{
example_pat

example_pat_failure_B
}
\keyword{datasets}
