% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaPurpleAir.R
\docType{data}
\name{example_raw_pas}
\alias{example_raw_pas}
\title{Example raw Purple Air Synoptic dataset}
\format{A tibble with 12657 rows and 32 columns of data.}
\source{
https://www.purpleair.com/json
}
\description{
The sample_pas dataset provides a quickly loadable version of
raw Purple Air syoptic data for practicing and code development. This
was generated by running \code{pas_load()} on 2019-03-14.

This dataset can be converted into a standard \emph{pas} dataset with:

\code{pas <- enhanceSynopticData(example_raw_pas)}
}
\seealso{
example_pas
}
\keyword{datasets}
