% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_sensor}
\alias{example_sensor}
\title{Example AirSensor Timeseries dataset}
\format{
An S3 object composed of "meta" and "data" data.
}
\usage{
example_sensor
}
\description{
The \code{example_sensor} dataset provides a quickly loadable version of
an \emph{airsensor} object for practicing and code examples.
This dataset was was generated on 2023-03-16 by running:

\preformatted{
library(AirSensor)

example_sensor <-
  pat_createNew(
    api_key = MY_API_READ_KEY,
    pas = example_pas,
    sensor_index = "9392",
    startdate = "2022-07-01",
    enddate = "2022-07-08",
    timezone = "UTC",
    verbose = TRUE
  ) %>%
  pat_createAirSensor(
    parameter = 'pm25',
    FUN = AirSensor::PurpleAirQC_hourly_AB_01
  )

save(example_sensor, file = "data/example_sensor.rda")
}
}
\keyword{datasets}
