% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_createNew.R
\name{pas_createNew}
\alias{pas_createNew}
\title{Create a new PurpleAir synoptic dataset}
\usage{
pas_createNew(
  api_key = NULL,
  countryCodes = NULL,
  stateCodes = NULL,
  counties = NULL,
  lookbackDays = 1,
  location_type = 0,
  read_keys = NULL,
  show_only = NULL,
  includePWFSL = TRUE,
  baseUrl = "https://api.purpleair.com/v1/sensors"
)
}
\arguments{
\item{api_key}{PurpleAir API Read Key.}

\item{countryCodes}{ISO 3166-1 alpha-2 country codes used to subset the data.
At least one countryCode must be specified.}

\item{stateCodes}{ISO-3166-2 alpha-2 state codes used to subset the data.
Specifying stateCodes is optional.}

\item{counties}{US county names or 5-digit FIPS codes used to subset the data.
Specifying counties is optional.}

\item{lookbackDays}{Number of days to "look back" for valid data. Data are
filtered to only include sensors with data more recent than \code{lookbackDays} ago.}

\item{location_type}{The \code{location_type} of the sensors. Possible values
are: 0 = Outside, 1 = Inside or \code{NULL} = both.}

\item{read_keys}{Optional comma separated list of sensor read_keys is required
for private devices. It is separate to the api_key and each sensor has its own
read_key. Submit multiple keys by separating them with a comma (,) character
for example: key-one,key-two,key-three.}

\item{show_only}{Optional comma separated list of sensor_index values. When
provided, the results are limited only to the sensors included in this list.}

\item{includePWFSL}{Logical specifying whether to calculate distances from 
PWFSL monitors.}

\item{baseUrl}{Base URL for the PurpleAir API.}
}
\value{
A PurpleAir Synoptic \emph{pas} object.
}
\description{
Download, parse and enhance synoptic data from PurpleAir and
return the results as a useful tibble with class \code{pa_synoptic}.

Steps include:

1) Download and parse synoptic data

2) Replace variable names with more consistent, more human readable names.

3) Add spatial metadata for each sensor including:
\itemize{
  \item{timezone -- olson timezone}
  \item{countryCode -- ISO 3166-1 alpha-2}
  \item{stateCode -- ISO 3166-2 alpha-2}
}

4) Convert data types from character to \code{POSIXct} and \code{numeric}.

Data requests are made with a bounding box determined from the required
\code{countryCodes} parameter. If a single country is specified and additional
\code{stateCodes} are specified, the bounding box will be limited to those
states. Withing a single state, \code{counties} may be used to further limit
the data.

If \code{show_only} is used to request specific sensors, the \code{countryCodes}
and \code{stateCodes} information is ignored when requesting data. But these 
are still used to help speed up the assignment of enhanced metadata..
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

initializeMazamaSpatialUtils()

pas <-
  pas_createNew(
    api_key = PURPLE_AIR_API_READ_KEY,
    countryCodes = "US",
    stateCodes = "CA",
    show_only = SCAQMD_SENSOR_INDICES,
    lookbackDays = 1,
    location_type = 0
  )

pas \%>\% pas_leaflet()

}, silent = FALSE)
}
}
\references{
\href{https://www2.purpleair.com}{PurpleAir}

\href{https://api.purpleair.com/}{PurpleAir API}

\href{https://www2.purpleair.com/policies/terms-of-service}{PurpleAir Terms of service}

\href{https://www2.purpleair.com/pages/license}{PurpleAir Data license}

\href{https://www2.purpleair.com/pages/attribution}{PurpleAir Data Attribution}
}
\seealso{
\link{pas_downloadParseRawData}

\link{pas_enhanceRawData}
}
