% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_esriMap.R
\name{pas_esriMap}
\alias{pas_esriMap}
\title{Create an ESRI Map of ws_monitor Object}
\usage{
pas_esriMap(pas, centerLon = NULL, centerLat = NULL,
  maptype = "worldStreetMap", mapRaster = NULL, zoom = NULL,
  showMap = FALSE)
}
\arguments{
\item{pas}{\emph{pa_synoptic} object}

\item{centerLon}{map center longitude}

\item{centerLat}{map center latitude}

\item{maptype}{map type}

\item{mapRaster}{optional RGB Raster* object returned from}

\item{zoom}{map zoom level}

\item{showMap}{option to show the ggplot if true}
}
\value{
Plots a map loaded from arcGIS REST with points for each monitor.
}
\description{
Creates an ESRI map of a \emph{pa_synoptic} object.

#' Available \code{maptype} options include:
\itemize{
\item{natGeo}
\item{worldStreetMap}
\item{worldTopoMap}
\item{satellite}
\item{deLorme}
}

Additional base maps are found at:
\url{http://resources.arcgis.com/en/help/arcgis-rest-api/index.html#/Basemaps/02r3000001mt000000/}

If \code{centerLon}, \code{centerMap} or \code{zoom} are not specified,
appropriate values will be calcualted using data from the
\code{pa_synoptic} dataframe.
}
\examples{
\dontrun{
pas <- pas_load()
CA <- pas \%>\% filter(stateCode == "CA")
pas_esriMap(CA, showMap=TRUE)
}
}
\seealso{
\code{\link{esriMap_plotOnStaticMap}}
}
\keyword{pa_synoptic}
