% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_filterArea.R
\name{pas_filterArea}
\alias{pas_filterArea}
\title{Area filtering for \emph{pas} objects}
\usage{
pas_filterArea(pas, w = NULL, e = NULL, s = NULL, n = NULL)
}
\arguments{
\item{pas}{\emph{pa_synoptic} dataframe}

\item{w}{west edge of bounding box (deg E)}

\item{e}{east edge of bounding box (deg E)}

\item{s}{south edge of bounding box (deg N)}

\item{n}{north edge of bounding box (deg N)}
}
\value{
A subset of the incoming \code{pas}.
}
\description{
Filters \emph{pa_synoptic} objects based on a bounding box.
}
\examples{
pas <- example_pas
range(pas$longitude)
range(pas$latitude)
scsb <- 
  pas \%>\%
  pas_filterArea(w = -118.10,
                 e = -118.07,
                 s = 33.75,
                 n = 33.78)
range(scsb$longitude)
range(scsb$latitude)

}
\seealso{
\link{pas_load}
}
