% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_filterNear.R
\name{pas_filterNear}
\alias{pas_filterNear}
\title{Find PurpleAir sensors within radial distance}
\usage{
pas_filterNear(pas = NULL, latitude = NULL, longitude = NULL,
  radius = "1 km")
}
\arguments{
\item{pas}{PurpleAir \emph{pas} object.}

\item{latitude}{a Target latitude.}

\item{longitude}{a Target longitude.}

\item{radius}{Distance from target with unit (i.e "15 km").}
}
\value{
A subset of the given \emph{pas} object.
}
\description{
Filter for Purple Air sensors within a specified distance from 
specified target coordinates.
}
\details{
\code{radius} Should be a numeric string with provided metric unit 
separated by a space, such as \code{"250 m"}.
}
\examples{
\dontrun{
# Near Diamond Bar, CA
diamond_bar <-
  pas_load() \%>\%
  pas_filterNear(latitude = 34.001667, 
                 longitude = -117.820833,
                 radius = "20 km")
                 
pas_leaflet(diamond_bar)
}
}
\seealso{
\link{pas_filter}

\link{pas_filterArea}
}
