% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_getColumn.R
\name{pas_getColumn}
\alias{pas_getColumn}
\title{Return column of data from filtered PurpleAir Synoptic objects}
\usage{
pas_getColumn(
  pas = NULL,
  name = NULL,
  states = NULL,
  pattern = ".*",
  isOutside = TRUE,
  isParent = TRUE
)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{name}{Name of the column to return}

\item{states}{Vector of recognized  ISO state codes}

\item{pattern}{Text pattern used to filter station labels}

\item{isOutside}{Logical, is the station located outside?}

\item{isParent}{Logigal, is the station a parent station?}
}
\value{
A column of data.
}
\description{
A filter for \emph{pas} objects to return the column of the 
stations of interest
}
