% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_getDeviceDeploymentIDs.R
\name{pas_getDeviceDeploymentIDs}
\alias{pas_getDeviceDeploymentIDs}
\title{Return timeseries identifiers from filtered PurpleAir Synoptic objects}
\usage{
pas_getDeviceDeploymentIDs(
  pas = NULL,
  states = PWFSLSmoke::US_52,
  pattern = ".*",
  isOutside = TRUE,
  isParent = TRUE
)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{states}{Vector of recognized ISO state codes.}

\item{pattern}{Text pattern used to filter station labels.}

\item{isOutside}{Logical, is the sensor located outside?}

\item{isParent}{Logigal, is the sensor a parent station?}
}
\value{
Vector of values.
}
\description{
The incoming \code{pas} object is first filtered based on the 
values of \code{states}, \code{patter}, \code{isOutside} and \code{isParent}.
The values associated with the \code{"deviceDeploymentID"} column are then 
returned.

This function is useful for returning a vector of unique time series
identifiers. These are used in the names of pre-generated \emph{pat} files
found in data archives.
}
\seealso{
\code{\link{pas_getColumn}},  \code{\link{pas_getLabels}}
}
