% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_getIDs.R
\name{pas_getIDs}
\alias{pas_getIDs}
\title{Return IDs from filtered PurpleAir Synoptic objects}
\usage{
pas_getIDs(pas = NULL, pattern = ".*", idPattern = ".*", isOutside = TRUE)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{pattern}{Text pattern used to filter station labels.}

\item{idPattern}{Text pattern used to filter \code{deviceDeploymentID}.}

\item{isOutside}{Logical, is the sensor located outside?}
}
\value{
Vector of values.
}
\description{
The incoming \code{pas} object is first filtered based on the 
values of \code{stateCodes}, \code{patter} and \code{isOutside}.
The values associated with the \code{"ID"} column are then returned.
}
\seealso{
\code{\link{pas_getColumn}},  \code{\link{pas_getLabels}}
}
