% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_getIDs.R
\name{pas_getIDs}
\alias{pas_getIDs}
\title{Return column of data from filtered PurpleAir Synoptic objects}
\usage{
pas_getIDs(pas = NULL)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{name}{Name of the column to return}
}
\value{
A column of data.
}
\description{
A filter for \emph{pas} objects to return the column of the 
stations of interest
}
