% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_getLabels.R
\name{pas_getLabels}
\alias{pas_getLabels}
\title{Return station labels from filtered PurpleAir Synoptic objects}
\usage{
pas_getLabels(pas = NULL, states = NULL, pattern = ".*",
  isOutside = TRUE, isParent = TRUE)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{states}{Vector of recognized  ISO state codes}

\item{pattern}{Text pattern used to filter station labels}

\item{isOutside}{Logical, is the station located outside?}

\item{isParent}{Logigal, is the station a parent station?}
}
\value{
A character vector of station labels.
}
\description{
A filter for \emph{pas} objects to return the labels of the 
stations of interest. Wrapper around \emph{pas_getColumn}.
}
