% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paSynoptic.R
\name{pas_hasSpatial}
\alias{pas_hasSpatial}
\title{Test for spatial metadata in \emph{pa_synoptic} object}
\usage{
pas_hasSpatial(pas)
}
\arguments{
\item{pas}{A \emph{pa_synoptic} object.}
}
\value{
\code{TRUE} if no data exist in \code{pas}, \code{FALSE} otherwise.
}
\description{
Convenience function for \code{nrow(pas) == 0}.
This makes for more readable code in functions that need to test for this.
}
\examples{
pas <- example_pas
pas_hasSpatial(pas)
}
