% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paSynoptic.R
\name{pas_isPas}
\alias{pas_isPas}
\title{Test for correct structure in a \emph{pa_synoptic} object}
\usage{
pas_isPas(pas = NULL)
}
\arguments{
\item{pas}{A \emph{pa_synoptic} object.}
}
\value{
\code{TRUE} if \code{pas} has the correct structure, \code{FALSE} otherwise.
}
\description{
The \code{pas} is checked for the "pas" class name
and presence of \strong{all} metadata columns:

\preformatted{
"ID",                               "label",
"DEVICE_LOCATIONTYPE",              "THINGSPEAK_PRIMARY_ID",
"THINGSPEAK_PRIMARY_ID_READ_KEY",   "THINGSPEAK_SECONDARY_ID",
"THINGSPEAK_SECONDARY_ID_READ_KEY", "latitude",
"longitude",                        "pm25",
"lastSeenDate",                     "sensorType",
"flag_hidden",                      "isOwner",
"humidity",                         "temperature",
"pressure",                         "age",
"parentID",                         "flag_highValue",
"flag_attenuation_hardware",        "Voc",
"Ozone1",                           "pm25_current",
"pm25_10min",                       "pm25_30min",
"pm25_1hr",                         "pm25_6hr",
"pm25_1day",                        "pm25_1week",
"statsLastModifiedDate",            "statsLastModifiedInterval",
"deviceID",                         "locationID",
"deviceDeploymentID",               "countryCode",
"stateCode",                        "timezone",
"airDistrict",                      "pwfsl_closestDistance",
"pwfsl_closestMonitorID",           "sensorManufacturer",
"targetPollutant",                  "technologyType",
"communityRegion"
}

The \code{pwfsl_}, "official", monitors are obtained from the USFS AirFire 
site using the \pkg{PWFSLSmoke} R package.
}
\examples{
pas_isPas(example_pas)
pas_isPas(1:10)
}
\keyword{pas}
