% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_leaflet.R
\name{pas_leaflet}
\alias{pas_leaflet}
\title{Leaflet interactive map of Purple Air sensors}
\usage{
pas_leaflet(pas, param = "pm25_1hr", radius = 10, opacity = 0.8,
  maptype = "terrain", outsideOnly = TRUE)
}
\arguments{
\item{pas}{Dnhanced dataframe of PurpleAir synoptic data.}

\item{param}{Value to plot -- defautls to \code{get('pm25_1hr')}.}

\item{radius}{Radius (pixels) of monitor circles.}

\item{opacity}{Opacity of monitor circles.}

\item{maptype}{Optional name of leaflet ProviderTiles to use, e.g. \code{terrain}.}

\item{outsideOnly}{Logical specifying subsetting for monitors marked as 'outside'.}
}
\value{
A leaflet "plot" object which, if not assigned, is rendered in 
Rstudio's 'Viewer' tab.
}
\description{
This function creates interactive maps that will be displayed in 
RStudio's 'Viewer' tab.

Typical usage would be to use the \code{param} argument to display pm25 
values from one of:
\itemize{
\item{"pm25_current"}
\item{"pm25_10min"}
\item{"pm25_30min"}
\item{"pm25_1hr"}
\item{"pm25_6hr"}
\item{"pm25_1day"}
\item{"pm25_1week"}
}
}
\details{
The \code{maptype} argument is mapped onto leaflet "ProviderTile" 
names. Current mappings include:
\enumerate{
\item{"roadmap"}{ -- "OpenStreetMap"}
\item{"satellite"}{ -- "Esri.WorldImagery"}
\item{"terrain"}{ -- ""Esri.WorldTopoMap"}
\item{"toner"}{ -- "Stamen.Toner"}
}

If a character string not listed above is provided, it will be used as the 
underlying map tile if available. See 
\url{https://leaflet-extras.github.io/leaflet-providers/} for a list of 
"provider tiles" to use as the background map.
}
\examples{
\dontrun{
# Methow Valley Clean Air Ambassador
initializeMazamaSpatialUtils()
pas <- pas_load()
mvcaa <- filter(pas, stringr::str_detect(pas$label, '^MV Clean Air'))
pas_leaflet(mvcaa, param="pm25_1hr")
}
}
