% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_load.R
\name{pas_load}
\alias{pas_load}
\title{Load Purple Air synoptic data}
\usage{
pas_load(datestamp = strftime(lubridate::today("America/Los_Angeles"),
  "\%Y\%m\%d"),
  baseUrl = "http://smoke.mazamascience.com/data/PurpleAir/pas/2019")
}
\arguments{
\item{datestamp}{A date string in ymd order.}

\item{baseUrl}{The base URL for synoptic data.}
}
\value{
A \code{pa_synoptic} dataframe.
}
\description{
A pre-generated \emph{pa_synoptic} dataframe will be loaded for
  the given date. These files are generated each day at 4am California time
  and provide a record of all currently installed Purple Air sensors for the
  day of interest.

  The \code{datestamp} can be anything that is understood by \code{lubrdiate::ymd()}
  including either of the following recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}

By default, today's date is used.
}
\examples{
\dontrun{
pas <- pas_load()
pas \%>\% 
  filter(stateCode == "CA") \%>\%
  pas_leaflet()
}
}
\seealso{
\link{pas_loadLatest}
}
