% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_palette.R
\name{pas_palette}
\alias{pas_palette}
\title{Color palettes for PurpleAir}
\usage{
pas_palette(pas, pal, param = NULL, ...)
}
\arguments{
\item{pas}{Enhanced data frame of PurpleAir synoptic data.}

\item{pal}{A predefined color palette. Can be of the following:
\itemize{
\item{"AQI"}
\item{"humidity}
\item{"temperature}
\item{"distance"}
}}
}
\value{
An object that consists of a label and color dataframe, and
calculated color values from PurpleAir sensors
}
\description{
Generates color palettes for PurpleAir synoptic data with the 
intention of having a reproducible functional color generator.
}
