% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_within.R
\name{pas_within}
\alias{pas_within}
\title{Find Purple Air sensors within radial distance}
\usage{
pas_within(pas, label = NULL, radius = "1 km")
}
\arguments{
\item{pas}{a pa_synoptic dataframe}

\item{label}{a target (centered) Purple Air sensor label}

\item{radius}{a distance with unit (i.e "15 km")}
}
\value{
a pa_synoptic object
}
\description{
Calculate Purple Air sensors within a specified distance from 
a target Purple Air sensor.
}
\details{
\code{radius} should be a numeric string with provided metric unit,
such as \code{"250 m"}.
}
\examples{
pas_within(example_pas,"SaratogaMiddleHighSchool", "1024 km")

}
