% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_aggregate.R
\name{pat_aggregate}
\alias{pat_aggregate}
\title{Time averages for PrupleAir time series}
\usage{
pat_aggregate(pat, period = "1 hour", stats = "mean",
  parameter = NULL, quickStats = TRUE)
}
\arguments{
\item{pat}{PurpleAir Timeseries "pat" object from \code{pat_loadLatest()}}

\item{period}{The time period to average to. Can be "sec", "min", "hour", 
"day", "DSTday", "week", "month", "quarter" or "year". A number can also
 precede these options followed by a space (i.e. "2 day" or "37 min").}

\item{stats}{The statistic to apply when aggregating the data; default is the 
mean. Can be one of "mean", "max", "min", "median", "count", 
"sd", "percentile", "tstats".}

\item{parameter}{The variable of timeseries data to aggergate, such as
"pm25_B", "temperature", etc.}

\item{quickStats}{a logical that if \code{TRUE} will override \code{stats} 
parameter and return and a data frame of "mean", "sd", and "count".}
}
\value{
Returns a dataframe with a date in class POSIXct.
}
\description{
Function to flexibly aggregate or expand data frames by 
different time periods and calculating vector-averages for a PurpleAir time 
series object. This function should be useful in many 
circumstances where it is necessary to work with different time average data.
}
\examples{
\dontrun{
# Hourly mean of Channel A PM2.5 density measurement
pat_aggregate(pat, "1 hour")
}

\dontrun{
# Maximum weekly temperature
pat \%>\%
  pat_aggregate("temperature", "1 week", stats = "max")
}
}
\keyword{pa_timeseries}
