% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_aggregate.R
\name{pat_aggregate}
\alias{pat_aggregate}
\title{Aggregate PurpleAir Timeseries objects}
\usage{
pat_aggregate(pat, FUN = mean, ...)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{FUN}{The function to be applied to each vector of numeric \code{pat} data.}

\item{...}{(optional) Additional arguments.}
}
\value{
Returns an aggregated \emph{pat} object.
}
\description{
Aggregate PurpleAir timeseries (\emph{pat}) objects along its 
datetime axis. Temporal aggregation involves splitting a \emph{pat} object into
seperate bins along its datetime axis. \code{FUN} is mapped to the \emph{pat}
numeric variables in each bin, and recombined back into an aggregated valid 
\emph{pat} object.
}
\details{
\code{FUN} must operate on univariate numeric vectors and return a 
scalar value. 

\code{...} Optional arguments or data can be explicitly provided for 
\code{FUN}. Additionally, advanced usage of \code{xts::split.xts} for 
custom aggregation is also valid.
}
\examples{
\dontrun{
# Standard hourly mean aggregation
avg <- pat_aggregate(pat, mean, na.rm = TRUE)

# Alternative 30 minute aggregation (advanced users only - see details.)
avg <- pat_aggregate(pat, f = 'minutes', k = 30, FUN = mean, na.rm = TRUE)
}
}
\keyword{pa_timeseries}
