% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_aggregateOutlierCounts.R
\name{pat_aggregateOutlierCounts}
\alias{pat_aggregateOutlierCounts}
\title{Aggregate data with count of outliers in each bin}
\usage{
pat_aggregateOutlierCounts(
  pat = NULL,
  period = "1 hour",
  windowSize = 23,
  thresholdMin = 8
)
}
\arguments{
\item{pat}{a PurpleAir Timeseries \emph{pat} object.}

\item{period}{time period to average over. Can be "sec", "min", "hour", 
"day", "DSTday", "week", "month", "quarter" or "year". A number can also
 precede these options followed by a space (i.e. "2 day" or "37 min").}

\item{windowSize}{the size of the rolling window}

\item{thresholdMin}{the minimum threshold value to detect outliers}
}
\value{
\code{data.frame} A data.frame with additional flag count vectors
}
\description{
Aggregate data with count of outliers in each bin
}
\examples{
\donttest{

df <- 
  pat_aggregateOutlierCounts(example_pat_failure_A)

library(ggplot2)
# Plot the counts 
multi_ggplot(
  # A Channel
  ggplot(df, aes(x = datetime, y = pm25_A_outlierCount)) + geom_point(),
  # B Channel
  ggplot(df, aes(x = datetime, y = pm25_B_outlierCount)) + geom_point(),
  # Humidity
  ggplot(df, aes(x = datetime, y = humidity_outlierCount)) + geom_point(),
  # Temperature 
  ggplot(df, aes(x = datetime, y = temperature_outlierCount)) + geom_point()
)

}

}
\seealso{
pat_aggregate
}
