% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_createAirSensor.R
\name{pat_createAirSensor}
\alias{pat_createAirSensor}
\title{Create an Air Sensor object}
\usage{
pat_createAirSensor(pat = NULL, period = "1 hour",
  parameter = "pm25", channel = "ab", min_count = 10, max_t = 100)
}
\arguments{
\item{pat}{PurpleAir Timeseries "pat" object}

\item{period}{The time period to average to. Can be "sec", "min", "hour", 
"day", "DSTday", "week", "month", "quarter" or "year". A number can also
 precede these options followed by a space (i.e. "2 day" or "37 min").}

\item{parameter}{parameter for which to create an \emph{as} object -- one of
"pm25", "humidity" or "temperature".}

\item{channel}{Data channel to use for PM2.5 -- one of "a", "b or "ab".}

\item{min_count}{Aggregation bins with fewer than `min_count` measurements
will be marked as `NA`.}

\item{max_t}{Aggregation bins where the two-sample t-test statistic is 
greater than `max_t` will be marked as `NA`. This applies only to
`parameter = "pm25"`.}
}
\value{
List with \code{meta} and \code{data} elements on a uniform time 
axis.

"as" object of aggregated time series PurpleAir data
}
\description{
Aggregates data from a \code{pat} object into an "Air Sensor" 
object that has appropriate metadata to be used with the *PWFSLSmoke* package.
}
\examples{
\dontrun{
as <- 
  example_pat \%>\%
  pat_filterDate(20180701, 20180901) \%>\%
  pat_createAirSensor(period = "1 hour")
}
}
\seealso{
\link{pat_createASTimeseries}
}
