% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_createASTimeseries.R
\name{pat_createASTimeseries}
\alias{pat_createASTimeseries}
\title{Create a AirSensor Timeseries object}
\usage{
pat_createASTimeseries(pat = NULL, period = "1 hour",
  pm25Source = "AB", qcType = NULL, removeOutliers = TRUE, ...)
}
\arguments{
\item{pat}{Purple Air timeseries data from \code{createPATimeseriesObject()}}

\item{period}{The time period to average to. Can be "sec", "min", "hour", 
"day". A number can also precede these options followed by a space 
(i.e. "15 min" or "1 hour").}

\item{pm25Source}{Which \code{pat} object columns to use when calculating
the "pm25" column: either \code{A}, \code{B} or \code{AB}}

\item{qcType}{named quality control method (currently none)}

\item{removeOutliers}{logical specifying whether to automatically remove
outliers using \code{\link{pat_outliers}} with default parameters.}

\item{...}{Extra parameters to fine tune time-averaging function}
}
\value{
List with \code{meta} and \code{data} elements on a uniform time 
axis.

"as_timeseries" list of aggregated time series PurpleAir data
}
\description{
Aggregates a PurpleAir timeseries object along the datetime
axis to produce a new dataframe period-aggregated timeseries data.
}
\examples{
\dontrun{
pat <- 
  AirSensor::example_pat \%>\%
  pat_filterDate(20180701, 20180901)
ast <- createASTimeseriesObject(pat, "1 hour")
}
}
