% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_createNew.R
\name{pat_createNew}
\alias{pat_createNew}
\title{Create a new PurpleAir timeseries dataset.}
\usage{
pat_createNew(
  api_key = NULL,
  pas = NULL,
  sensor_index = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = "UTC",
  average = 0,
  fields = SENSOR_HISTORY_PM25_FIELDS,
  baseUrl = "https://api.purpleair.com/v1/sensors",
  verbose = FALSE
)
}
\arguments{
\item{api_key}{PurpleAir API Read Key. If \code{api_key = NULL}, it
will be obtained using \code{getAPIKey("PurpleAir-read")}.
See \code{MazamaCoreUtils::\link[MazamaCoreUtils:setAPIKey]{setAPIKey}}.}

\item{pas}{Previously generated \emph{pas} object containing \code{sensor_index}.}

\item{sensor_index}{PurpleAir sensor unique identifier.}

\item{startdate}{Desired start time (ISO 8601) or \code{POSIXct}.}

\item{enddate}{Desired end time (ISO 8601) or \code{POSIXct}.}

\item{timezone}{Olson timezone used to interpret dates.}

\item{average}{Temporal averaging in minutes performed by PurpleAir. One of:
0 (raw), 10, 30, 60 (hour), 360, 1440 (day).}

\item{fields}{Character string with PurpleAir field names for the Get Sensor Data API.}

\item{baseUrl}{Base URL for the PurpleAir API.}

\item{verbose}{Logical controlling the generation of warning and error messages.}
}
\value{
A PurpleAir Timeseries \emph{pat} object.
}
\description{
Cerate a \code{pat} object for a specific \code{sensor_index}.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor)

pat <-
  pat_createNew(
    api_key = MY_API_READ_KEY,
    pas = MY_PAS,
    sensor_index = "10168",
    startdate = "2023-01-01",
    enddate = "2023-01-08",
    timezone = "UTC",
    verbose = TRUE
  )

View(pat$meta[1:100,])

}, silent = FALSE)
}
}
\references{
\href{https://www2.purpleair.com}{PurpleAir}

\href{https://api.purpleair.com}{PurpleAir API}

\href{https://www2.purpleair.com/policies/terms-of-service}{PurpleAir Terms of service}

\href{https://www2.purpleair.com/pages/license}{PurpleAir Data license}

\href{https://www2.purpleair.com/pages/attribution}{PurpleAir Data Attribution}
}
