% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_createNew.R
\name{pat_createNew}
\alias{pat_createNew}
\title{Load latest PurpleAir time series data}
\usage{
pat_createNew(pas = NULL, label = NULL, id = NULL,
  startdate = NULL, enddate = NULL, timezone = NULL,
  baseURL = "https://api.thingspeak.com/channels/")
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{label}{PurpleAir sensor 'label'.}

\item{id}{PurpleAir sensor 'ID'.}

\item{startdate}{Desired UTC start time (ISO 8601).}

\item{enddate}{Desired UTC end time (ISO 8601).}

\item{timezone}{Timezone used to interpret start and end dates.}

\item{baseURL}{Base URL for Thingspeak API.}
}
\value{
A PurpleAir Timeseries \emph{pat} object.
}
\description{
Retrieve and parse timeseries data from the Thingspeak API for 
specific PurpleAir sensors.
}
\note{
When \code{timezone = NULL}, the default, dates are interpreted to be 
in the local timezone for the sensor of interest.
}
\examples{
\donttest{
setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
pas <- pas_load()
pat <- pat_createNew(pas, "Seattle", startdate = 20180701, enddate = 20180901)
pat_multiplot(pat)
}
}
\seealso{
\link{downloadParseTimeseriesData}
}
