% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_dailySoHIndex.R
\name{pat_dailySoHIndex}
\alias{pat_dailySoHIndex}
\title{Daily State of Health indexing}
\usage{
pat_dailySoHIndex(pat = NULL)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}
}
\description{
This function calculates the \code{pat_dailySoH} function and 
returns a tibble containing a state of health index for each day of the 
\emph{pat} provided. The tibble contains a \emph{datetime}, \emph{SoH_index}, 
and \emph{SoH_index_bin}. The \emph{SoH_index} is a value normalized between 
0 and 1 where 0 represents low confidence in the sensor data and 1 represents 
high confidence. The \emph{SoH_index_bin} is either 0, 1, or 2 representing
poor, fair, and good data respectively.
}
\examples{
 
tbl <- 
  example_pat_failure_A \%>\%
  pat_dailySoHIndex() 
  


}
