% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_dailySoHIndex_00.R
\name{pat_dailySoHIndex_00}
\alias{pat_dailySoHIndex_00}
\title{State of Health index plot}
\usage{
pat_dailySoHIndex_00(
  pat = NULL,
  minPctReporting = 50,
  breaks = c(0, 0.2, 0.8, 1)
)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{minPctReporting}{Percent reporting threshold for A and B channels.}

\item{breaks}{Breaks used to convert index values into index bins.}
}
\description{

}
\examples{
 
library(AirSensor)

tbl <- 
  example_pat_failure_A \%>\%
  pat_dailySoHIndex_00() 
  
head(tbl)
  
}
