% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_dailyStateOfHealth.R
\name{pat_dailyStateOfHealth}
\alias{pat_dailyStateOfHealth}
\title{Daily state of health}
\usage{
pat_dailyStateOfHealth(pat = NULL,
  SoH_functions = c("PurpleAirSoH_dailyPctDC",
  "PurpleAirSoH_dailyPctReporting", "PurpleAirSoH_dailyPctValid",
  "PurpleAirSoH_dailyCorrelation"))
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{SoH_functions}{Vector of function names. All the passed in functions 
must output tibbles with a daily `datetime` variable and must cover the same
period of time.}
}
\description{
This function combines the output of the State of Health (SoH) 
function arguments into a single tibble.
}
\examples{
 
SoH <- 
  example_pat_failure_B \%>\%
  pat_dailyStateOfHealth() 
head(SoH)
#timeseriesTbl_multiplot(tbl, ylim = c(0,100))
}
