% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paTimeseries.R
\name{pat_distinct}
\alias{pat_distinct}
\title{Retain only distinct data records in pat$data}
\usage{
pat_distinct(pat)
}
\arguments{
\item{pat}{\emph{pat} object}
}
\value{
A \emph{pat} object with no duplicated data records.
}
\description{
Convenience wrapper for 
\code{pat$data <- dplyr::distinct(pat$data)}.
}
