% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_dygraph.R
\name{pat_dygraph}
\alias{pat_dygraph}
\title{Create Interactive Time Series Plot}
\usage{
pat_dygraph(pat, plottype = NULL, title = NULL, xlab = NULL,
  ylab = NULL, tlim = NULL, rollPeriod = 1, showLegend = TRUE,
  colors = NULL)
}
\arguments{
\item{pat}{Purple Air Timeseries "pat" object from \code{createPATimeseriesObject()}}

\item{plottype}{Quick-reference plot types: "pm25", "humidity", "temperature"}

\item{title}{title text}

\item{xlab}{optional title for the x axis}

\item{ylab}{optional title for the y axis}

\item{tlim}{optional vector with start and end times (integer or character
representing YYYYMMDD[HH])}

\item{rollPeriod}{rolling mean to be applied to the data}

\item{showLegend}{logical to toggle display of the legend}

\item{colors}{string vector of colors to be used for plotting}
}
\value{
Initiates the interactive dygraph plot in RStudio's 'Viewer' tab.
}
\description{
This function creates interactive graphs that will be displayed
in RStudio's 'Viewer' tab.
}
\examples{
\dontrun{
pas <- example_pas
nb <- pat_load(pas, "North Bend Weather", startdate = 20180801, enddate = 20180901)
subset_nb <- pat_sample(pat=nb, sampleSize = 1000, setSeed = 1)
pat_dygraph(pat = subset_nb, xlab = "2018", rollPeriod = 7)
}
}
\keyword{pa_timeseries}
