% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_externalFit.R
\name{pat_externalFit}
\alias{pat_externalFit}
\title{Linear model fitting of PurpleAir and federal PWFSL time series data}
\usage{
pat_externalFit(pat = NULL, replaceOutliers = TRUE, showPlot = TRUE,
  size = 1, lm_shape = 15, ts_shape = 1, color = "purple",
  alpha = 0.5, xylim = NULL, channel = "ab",
  qc_algorithm = "hourly_AB_01", min_count = 10, pa_color = "purple",
  pwfsl_color = "black")
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{replaceOutliers}{Logical specifying whether or not to replace outliers.}

\item{showPlot}{Logical specifying whether to generate a model fit plot.}

\item{size}{Size of points.}

\item{lm_shape}{Symbol to use for linear model points.}

\item{ts_shape}{Symbol to use for time series plot points.}

\item{color}{Color of points.}

\item{alpha}{Opacity of points.}

\item{xylim}{Vector of (lo,hi) limits used as limits on the correlation plot 
axes -- useful for zooming in.}

\item{channel}{Data channel to use for PM2.5 -- one of "a", "b or "ab".}

\item{qc_algorithm}{Named QC algorithm to apply to hourly aggregation stats.}

\item{min_count}{Aggregation bins with fewer than `min_count` measurements
will be marked as `NA`.}

\item{pa_color}{Color of hourly points}

\item{pwfsl_color}{Color of hourly points}
}
\value{
A linear model, fitting the `pat` PurpleAir readings to the closest
PWFSL monitor readings.
}
\description{
Produces a linear model between data from PurpleAir and data 
from the closest PWFSL monitor.

A diagnostic plot is produced if `showPlot = TRUE`.
}
\examples{
\dontrun{
pat <- pat_load("SCAH_22")
pat_externalFit(pat)
}
}
