% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paTimeseries.R
\name{pat_extractDataFrame}
\alias{pat_extractDataFrame}
\alias{pat_extractData}
\alias{pat_extractMeta}
\title{Extract dataframes from \emph{pat} objects}
\usage{
pat_extractData(pat)

pat_extractMeta(pat)
}
\arguments{
\item{pat}{\emph{pat} object to extract dataframe from.}
}
\value{
A dataframe from the given \emph{pat} object
}
\description{
These functions are convenient wrappers for extracting the dataframes that
comprise a \emph{pat} object. These functions are designed to be useful when
manipulating data in a pipeline chain using \code{\%>\%}.

Below is a table showing equivalent operations for each function.

\tabular{ll}{
  \strong{Function} \tab \strong{Equivalent Operation}\cr
  \code{pat_extractData(pat)} \tab \code{pat[["data"]]}\cr
  \code{pat_extractMeta(pat)} \tab \code{pat[["meta"]]}
}
}
