% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_filterData.R
\name{pat_filterData}
\alias{pat_filterData}
\title{General purpose filtering for \emph{pat} objects}
\usage{
pat_filterData(pat = NULL, ...)
}
\arguments{
\item{pat}{\emph{pa_timeseries} dataframe}

\item{...}{logical predicates defined in terms of the variables in the 
\code{pat$data}.}
}
\value{
A subset of the incoming \code{pat}.
}
\description{
A generalized data filter for \emph{pat} objects to 
choose rows/cases where conditions are true.  Multiple conditions are 
combined with \code{&} or seperated by a comma. Only rows where the condition 
evaluates to TRUE are kept.Rows where the condition evaluates to \code{NA}
aredropped.
}
\examples{
\dontrun{
unhealthy <- pat_filterData(pat, pm25_A > 55.5, pm25_B > 55.5) 
}
}
\seealso{
\link{pat_filterDate}
}
