% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_internalFit.R
\name{pat_internalFit}
\alias{pat_internalFit}
\title{Linear model fitting of channel A and B data}
\usage{
pat_internalFit(pat = NULL, showPlot = TRUE, size = 1, shape = 15,
  color = "purple", alpha = 0.25, xylim = NULL)
}
\arguments{
\item{pat}{Purple Air Timeseries "pat" object}

\item{showPlot}{logical specifying whether to generate a model fit plot}

\item{size}{size of points}

\item{shape}{symbol to use for points}

\item{color}{color of points}

\item{alpha}{opacity of points}

\item{xylim}{vector of (lo,hi) limits used as limits on the correlation plot 
axes -- useful for zooming in}
}
\value{
A linear model, fitting the `pat` B channel readings to A channel 
readings.
}
\description{
Uses a liner model to fit data from channel B to data from 
channel A.

A diagnostic plot is produced if `showPlot = TRUE`.
}
