% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_internalFit.R
\name{pat_internalFit}
\alias{pat_internalFit}
\title{Return a linear model gitting channel A and B}
\usage{
pat_internalFit(pat, subset = NULL, weights = NULL, showPlot = TRUE)
}
\arguments{
\item{pat}{Purple Air Timeseries "pat" object}

\item{subset}{(From `stats::lm`) An optional vector specifying a subset of
observations to be used in the fitting process.}

\item{weights}{(From `stats::lm`) An optional vector of weights to be used in
the fitting process. Should be NULL or a numeric vector. If non-NULL,
weighted least squares is used with weights weights (that is, minimizing
sum(w*e^2)); otherwise ordinary least squares is used.}

\item{showPlot}{logical specifying whether to generate a model fit plot}
}
\value{
A linear model, fitting the `pat` B channel readings to A channel readings.
}
\description{
Uses a liner model to fit data from channel B to data from 
channel A.

A diagnostic plot is produced if `showPlot = TRUE`.
}
