% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paTimeseries.R
\name{pat_isPat}
\alias{pat_isPat}
\title{Test for correct structure in a \emph{pat} object}
\usage{
pat_isPat(pat = NULL)
}
\arguments{
\item{pat}{\emph{pat} object}
}
\value{
\code{TRUE} if \code{pat} has the correct structure, \code{FALSE} otherwise.
}
\description{
The \code{pat} is checked for the 'pat' class name
and presence of core metadata columns:
\itemize{
  \item{ID -- Purple Air ID}
  \item{label -- location label}
  \item{sensorType -- Purple Air sensor type}
  \item{longitude -- decimal degrees E}
  \item{latitude -- decimal degrees N}
  \item{timezone -- Olson timezone}
  \item{countryCode -- ISO 3166-1 alpha-2}
  \item{stateCode -- ISO 3166-2 alpha-2}
  \item{pwfsl_closestDistance -- distance in meters from an official monitor}
  \item{pwfsl_closestMonitorID -- identifer for the nearest official monitor}
}

The "pwfsl", official, monitors are obtained from the USFS AirFire site 
using the \pkg{PWFSLSmoke} R package.
}
\examples{
pat_isPat(example_pat)

}
\keyword{pat}
