% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_join.R
\name{pat_join}
\alias{pat_join}
\title{Join PurpleAir time series data}
\usage{
pat_join(..., param = NULL)
}
\arguments{
\item{...}{Any number of valid \code{pat} objects}

\item{param}{Not used.}
}
\value{
\code{pat} time series object
}
\description{
Create a union of any number of \code{pat} objects along a 
non-continuous datetime axis.
}
\examples{
\dontrun{
jul01_15 <- 
  pat \%>\%
  pat_filterDate(20180701, 20180715)

aug09_22 <- 
  pat \%>\% 
  pat_filterDate(20180809, 20180822)
  
x <- pat_join(jul01_15, aug09_22)
}

}
