% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_load.R
\name{pat_load}
\alias{pat_load}
\title{Load Purple Air Timeseries data}
\usage{
pat_load(label = NULL, startdate = NULL, enddate = NULL, days = 7,
  timezone = "America/Los_Angeles",
  baseUrl = "http://smoke.mazamascience.com/data/PurpleAir/pat")
}
\arguments{
\item{label}{Purple Air 'label'}

\item{startdate}{desired start datetime (ISO 8601)}

\item{enddate}{desired end datetime (ISO 8601)}

\item{days}{number of days of data to include}

\item{timezone}{Timezone used to interpret datestamp.}

\item{baseUrl}{Base URL for synoptic data.}
}
\value{
An object of class "pa_timeseries".
}
\description{
A pre-generated \code{pa_timeseries} object will be loaded for
the given time interval. Archived data for SCAQMD sensors go back to 
January, 2018.

Dates can be anything that is understood by 
\code{lubrdiate::parse_date_time()} including either of the following 
recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}

By default, the current week is loaded.
}
\examples{
\dontrun{
pat <- pat_load("SCNP_20", 20190411, 20190521)
pat_multiplot(pat)
}
}
\seealso{
\link{pat_loadMonth}

\link{.dateRange}
}
