% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_load.R
\name{pat_load}
\alias{pat_load}
\title{Load PurpleAir time series data for a time period}
\usage{
pat_load(
  label = NULL,
  startdate = NULL,
  enddate = NULL,
  days = 7,
  timezone = "America/Los_Angeles"
)
}
\arguments{
\item{label}{PurpleAir sensor 'label'.}

\item{startdate}{Desired start time (ISO 8601).}

\item{enddate}{Desired end time (ISO 8601).}

\item{days}{Number of days of data to include.}

\item{timezone}{Timezone used to interpret start and end dates.}
}
\value{
A PurpleAir Timeseries \emph{pat} object.
}
\description{
A pre-generated PurpleAir Timeseries \emph{pat} object will be 
loaded for the given time interval. Archived data for SCAQMD sensors go back 
to January, 2018.

Dates can be anything that is understood by 
\code{lubrdiate::parse_date_time()} including either of the following 
recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}

By default, the current week is loaded.
}
\examples{
\donttest{
setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
pat <- pat_load("SCNP_20", 20190411, 20190521)
pat_multiplot(pat)
}
}
\seealso{
\link{pat_loadLatest}

\link{pat_loadMonth}

\link{pat_createNew}
}
