% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_load.R
\name{pat_load}
\alias{pat_load}
\title{Get Purple Air timeseries data}
\usage{
pat_load(pas = NULL, name = NULL, id = NULL, startdate = NULL,
  enddate = NULL, baseURL = "https://api.thingspeak.com/channels/")
}
\arguments{
\item{pas}{Purple Air 'enhanced' synoptic data}

\item{name}{Purple Air 'label'}

\item{id}{Purple Air 'ID'}

\item{startdate}{desired start datetime (ISO 8601)}

\item{enddate}{desired end datetime (ISO 8601)}

\item{baseURL}{Base URL for Thingspeak API}
}
\value{
List with \code{meta} and \code{data} elements
}
\description{
Retrieve and parse timeseries data from the Thingspeak API for 
specific PurpleAir sensors.
}
\note{
Dates are interpreted to be in the local timzone for the sensor of 
interest.
}
\examples{
\dontrun{
initializeMazamaSpatialUtils()
pas <- pas_load()
pat <- pat_load(pas, "Seattle", startdate = 20180701, enddate = 20180901)
}
}
\seealso{
\link{downloadParseTimeseriesData}
}
