% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_load.R
\name{pat_load}
\alias{pat_load}
\title{Load PurpleAir time series data for a time period}
\usage{
pat_load(
  pas = NULL,
  label = NULL,
  id = NULL,
  startdate = NULL,
  enddate = NULL,
  days = 7,
  timezone = "America/Los_Angeles"
)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{label}{PurpleAir sensor 'label'.}

\item{id}{PurpleAir sensor 'ID'.}

\item{startdate}{Desired start time (ISO 8601).}

\item{enddate}{Desired end time (ISO 8601).}

\item{days}{Number of days of data to include.}

\item{timezone}{Timezone used to interpret start and end dates.}
}
\value{
A PurpleAir Timeseries \emph{pat} object.
}
\description{
A pre-generated PurpleAir Timeseries \emph{pat} object will be 
loaded for the given time interval if available. Data are 
loaded from the archive set with either \code{setArchiveBaseUrl()} or 
\code{setArchiveBaseDir()} for locally archived files.

Dates can be anything that is understood by 
\code{lubrdiate::parse_date_time()} including either of the following 
recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}
}
\note{
Starting with \pkg{AirSensor} version 0.6, archive file names are 
generated with a "sensor-deployment" identifier by combining a unique 
location ID with a unique device ID. These "sensor-deployment" identifiers 
guarantee that movement of a sensor will result in the creation of a new
time series.
}
\examples{
\donttest{
# TODO:  This needs to be updated to use USFS data
setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
pat <- pat_load("SCNP_20", 20190411, 20190521)
pat_multiplot(pat)
}
}
\seealso{
\link{pat_loadLatest}

\link{pat_loadMonth}

\link{pat_createNew}
}
