% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_loadLatest.R
\name{pat_loadLatest}
\alias{pat_loadLatest}
\title{Load latest PurpleAir time series data}
\usage{
pat_loadLatest(pas = NULL, name = NULL, id = NULL,
  startdate = NULL, enddate = NULL,
  baseURL = "https://api.thingspeak.com/channels/")
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{name}{PurpleAir sensor 'label'.}

\item{id}{PurpleAir sensor 'ID'.}

\item{startdate}{Desired start datetime (ISO 8601).}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{baseURL}{Base URL for Thingspeak API.}
}
\value{
A PurpleAir Timeseries \emph{pat} object.
}
\description{
Retrieve and parse timeseries data from the Thingspeak API for 
specific PurpleAir sensors.
}
\note{
Dates are interpreted to be in the local timzone for the sensor of 
interest.
}
\examples{
\donttest{
pas <- pas_load()
pat <- pat_loadLatest(pas, "Seattle", startdate = 20180701, enddate = 20180901)
pat_multiplot(pat)
}
}
\seealso{
\link{downloadParseTimeseriesData}
}
