% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_loadLatest.R
\name{pat_loadLatest}
\alias{pat_loadLatest}
\title{Load PurpleAir time series data for a week}
\usage{
pat_loadLatest(label = NULL)
}
\arguments{
\item{label}{PurpleAir sensor 'label'.}
}
\value{
A PurpleAir Timeseries \emph{pat} object.
}
\description{
A pre-generated PurpleAir Timeseries \emph{pat} object will be 
loaded containing data for the most recent 7-day interval.
}
\examples{
\donttest{
setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
pat <- pat_loadLatest("SCNP_20")
pat_multiplot(pat)
}
}
\seealso{
\link{pat_load}

\link{pat_loadMonth}

\link{pat_createNew}
}
