% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_loadLatest.R
\name{pat_loadLatest}
\alias{pat_loadLatest}
\title{Load PurpleAir time series data for a week}
\usage{
pat_loadLatest(pas = NULL, label = NULL, id = NULL)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{label}{PurpleAir sensor 'label'.}

\item{id}{PurpleAir sensor 'ID'.}
}
\value{
A PurpleAir Timeseries \emph{pat} object.
}
\description{
A pre-generated PurpleAir Timeseries \emph{pat} object will be 
loaded containing data for the most recent 7-day interval. Data are loaded
from the archive set with either \code{setArchiveBaseUrl()} or 
\code{setArchiveBaseDir()} for locally archived files.
}
\note{
Starting with \pkg{AirSensor} version 0.6, archive file names are 
generated with a "sensor-deployment" identifier by combining a unique 
location ID with a unique device ID. These "sensor-deployment" identifiers 
guarantee that movement of a sensor will result in the creation of a new
time series.
}
\examples{
\donttest{
# TODO:  This needs to be updated to use USFS data
setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
pat <- pat_loadLatest("SCNP_20")
pat_multiplot(pat)
}
}
\seealso{
\link{pat_load}

\link{pat_loadMonth}

\link{pat_createNew}
}
