% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_loadMonth.R
\name{pat_loadMonth}
\alias{pat_loadMonth}
\title{Load PurpleAir time series data for a month}
\usage{
pat_loadMonth(label = NULL, datestamp = NULL,
  timezone = "America/Los_Angeles")
}
\arguments{
\item{label}{Purple Air sensor 'label'}

\item{datestamp}{Date string in ymd order.}

\item{timezone}{Timezone used to interpret \code{datestamp}.}
}
\value{
A PurpleAir Timeseries \emph{pat} object.
}
\description{
A pre-generated PurpleAir Timeseries \emph{pat} object will be 
loaded for the given month. Archived data for SCAQMD sensors go back to 
January, 2018.

The \code{datestamp} must be in the following format:

\itemize{
\item{\code{"YYYYmm"}}
}

By default, the current month is loaded.
}
\examples{
\donttest{
setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
may <- pat_loadMonth("SCNP_20", 201905)
pat_multiplot(may)
}
}
\seealso{
\link{pat_load}

\link{pat_loadLatest}

\link{pat_createNew}
}
