% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_loadMonth.R
\name{pat_loadMonth}
\alias{pat_loadMonth}
\title{Load PurpleAir time series data for a month}
\usage{
pat_loadMonth(
  pas = NULL,
  label = NULL,
  id = NULL,
  datestamp = NULL,
  timezone = "America/Los_Angeles"
)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{label}{PurpleAir sensor 'label'.}

\item{id}{PurpleAir sensor 'ID'.}

\item{datestamp}{Date string in ymd order.}

\item{timezone}{Timezone used to interpret \code{datestamp}.
\code{make.names(label)} when assembilng the file path.}
}
\value{
A PurpleAir Timeseries \emph{pat} object.
}
\description{
A pre-generated PurpleAir Timeseries \emph{pat} object will be 
loaded for the month requested with \code{datestamp} if available. Data are 
loaded from the archive set with either \code{setArchiveBaseUrl()} or 
\code{setArchiveBaseDir()} for locally archived files.

The \code{datestamp} must be in the following format:

\itemize{
\item{\code{"YYYYmm"}}
}

By default, the current month is loaded.
}
\note{
Starting with \pkg{AirSensor} version 0.6, archive file names are 
generated with a "sensor-deployment" identifier by combining a unique 
location ID with a unique device ID. These "sensor-deployment" identifiers 
guarantee that movement of a sensor will result in the creation of a new
time series.
}
\examples{
\donttest{
# TODO:  This needs to be updated to use USFS data
setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
may <- pat_loadMonth("SCNP_20", 201905)
pat_multiplot(may)
}
}
\seealso{
\link{pat_load}

\link{pat_loadLatest}

\link{pat_createNew}
}
