% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_monitorComparison.R
\name{pat_monitorComparison}
\alias{pat_monitorComparison}
\title{Comparison of Purple Air and federal donitoring data}
\usage{
pat_monitorComparison(pat = NULL, ylim = NULL,
  replaceOutliers = TRUE, a_size = 1, a_shape = 15,
  a_color = "gray80", b_size = 1, b_shape = 15, b_color = "gray80",
  ab_alpha = 0.5, hourly_size = 2, hourly_shape = 1,
  hourly_stroke = 0.6, pa_color = "purple", pwfsl_color = "black")
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{ylim}{Vector of (lo,hi) y-axis limits used as limits.}

\item{replaceOutliers}{Logical specifying whether replace outliers in the
\emph{pat} object.}

\item{a_size}{Size of pm25_A points.}

\item{a_shape}{Symbol to use for pm25_A points.}

\item{a_color}{Color of pm25_A points.}

\item{b_size}{Size of pm25_B points.}

\item{b_shape}{Symbol to use for pm25_B points.}

\item{b_color}{Color of pm25_B points.}

\item{ab_alpha}{Opacity of pm25_A, pm25_B points.}

\item{hourly_size}{Size of hourly points.}

\item{hourly_shape}{Symbol to use for hourly points}

\item{hourly_stroke}{Line width of hourly points}

\item{pa_color}{Color of hourly points}

\item{pwfsl_color}{Color of hourly points}
}
\value{
A ggplot object.
}
\description{
Creates and regurns a ggplot object that plots raw \emph{pat}
data, hourly aggregated \emph{pat} data and hourly data from the nearest
federal monitor from the PWFSL database.
}
\examples{
\dontrun{
pat_monitorComparison(example_pat)
}
}
