% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_scatterplot.R
\name{pat_scatterplot}
\alias{pat_scatterplot}
\title{Compare all variables in a Purple Air Timeseries object}
\usage{
pat_scatterplot(pat, parameters = c("datetime", "pm25_A", "pm25_B",
  "temperature", "humidity"))
}
\arguments{
\item{pat}{Purple Air Timeseries \emph{pat} object}

\item{parameters}{vector of parameters to include}
}
\value{
Tibble portion of the \emph{pat} object, subset.
}
\description{
Creates a multi-panel scatterplot comparing all variables in the
\emph{pat} object. If any variables have no valid data, they are omitted from
the plot.

The list of available parameters include:

\itemize{
\item{\code{datetime} -- measurement time}
\item{\code{pm25_A} -- A channel PM2.5 (ug/m3)}
\item{\code{pm25_A} -- A channel PM2.5 (ug/m3)}
\item{\code{temperature} -- temperature (F)}
\item{\code{humidity} -- humidity (\%)}
\item{\code{uptime} -- seconds since last reset}
\item{\code{acd0} -- analog input voltage}
\item{\code{rssi} -- wifi signal strength (dBm)}
}
}
