% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAirQC_aggregationPlot.R
\name{PurpleAirQC_aggregationPlot}
\alias{PurpleAirQC_aggregationPlot}
\title{Plot the output from the pat_aggregateOutlierCounts() function}
\usage{
PurpleAirQC_aggregationPlot(aggregationStats = NULL,
  parameterGroup = "all", parameters = NULL, ylim = "fixed")
}
\arguments{
\item{aggregationStats}{PurpleAir Timeseries \emph{aggregationStats} object.}

\item{parameterGroup}{Quick-reference plot types based on a group of 
parameters: "all", "humidity", "pm25_A", "pm25_B", or "temperature"}

\item{parameters}{Custom vector of aggregation parameters to view}

\item{ylim}{Either "free_y", which scales the y axis automatically for each 
plot or "fixed" where the y limits of each plot are identical}
}
\description{
A plotting function that uses ggplot2 to display a plot of each 
output category from the pat_aggregateOutlierCounts() function. Created to 
have a quick look at all the stats to help identify necessary quality control
methods for PurpleAir Timeseries \emph{aggregationStats} objects.
}
\examples{
\dontrun{
nb <- pat_createNew(example_pas, 
                    "North Bend Weather", 
                    startdate = 20180801, 
                    enddate = 20180901)
aggregationStats <- pat_aggregateOutlierCounts(nb)
PurpleAirQC_aggregationPlot(aggregationStats, 
                            parameterGroup = "humidity", 
                            ylim = "free_y")
}

}
