% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAirQC_hourly_AB_01.R
\name{PurpleAirQC_hourly_AB_01}
\alias{PurpleAirQC_hourly_AB_01}
\title{Apply hourly aggregation QC using "AB_O1" algorithm}
\usage{
PurpleAirQC_hourly_AB_01(pat = NULL, min_count = 20, returnAllColumns = FALSE)
}
\arguments{
\item{pat}{A PurpleAir timeseries object.}

\item{min_count}{Aggregation bins with fewer than \code{min_count} measurements
will be marked as \code{NA}.}

\item{returnAllColumns}{Logical specifying whether to return all columns
of statistical data generated for QC algorithm or just the final \code{pm25}
result.}
}
\value{
Data frame with columns \code{datetime} and \code{pm25}.
}
\description{
Creates a \code{pm25} timeseries by averaging aggregated data
from the A and B channels and applying the following QC logic:

\enumerate{
\item{Create pm25 by averaging the A and B channel aggregation means}
\item{Invalidate data where:  (min_count < 20)}
\item{Invalidate data where:  (p-value < 1e-4) & (mean_diff > 10)}
\item{Invalidate data where:  (pm25 < 100) & (mean_diff > 20)}
}
}
\note{
Purple Air II sensors reporting after the June, 2019 firmware
upgrade report data every 2 minutes or 30 measurements per hour. The default
setting of \code{min_count = 20} is equivalent to a required data recovery
rate of 67%.
}
