% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwfsl_load.R
\name{pwfsl_loadLatest}
\alias{pwfsl_loadLatest}
\title{Get PWFSLSmoke Monitoring Data}
\usage{
pwfsl_loadLatest()
}
\value{
List with \code{meta} and \code{data} elements
}
\description{
Loads recent PM2.5 monitoring data from the US Forest Service
Pacific Wildland Fire Sciences Lab.

This function is a wrapper around \code{PWFSLSmoke::monitor_loadLatest}.
}
\examples{
\dontrun{
pwfsl <- pwfsl_load()
}
}
