% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_load.R
\name{sensor_load}
\alias{sensor_load}
\title{Load hourly-aggregated Purple Air data}
\usage{
sensor_load(collection = "scaqmd", startdate = NULL, enddate = NULL,
  days = 7, timezone = "America/Los_Angeles",
  baseUrl = "http://smoke.mazamascience.com/data/PurpleAir/airsensor")
}
\arguments{
\item{collection}{Name associated with the collection.}

\item{startdate}{Desired start datetime (ISO 8601).}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{days}{Number of days of data to include.}

\item{timezone}{Timezone used to interpret datestamp.}

\item{baseUrl}{Base URL for the \code{airsensor} data archive.}
}
\value{
An object of class "airsensor".
}
\description{
A pre-generated \code{airsensor} object will be loaded for
the given time interval. Archived data for SCAQMD sensors go back to 
January, 2018.

Dates can be anything that is understood by 
\code{lubrdiate::parse_date_time()} including either of the following 
recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}

By default, the current week is loaded.
}
\examples{
\donttest{
sensor_load("scaqmd", 20190411, 20190521) \%>\%
  PWFSLSmoke::monitor_timeseriesPlot(style = 'gnats')
}
}
\seealso{
\link{sensor_loadMonth}
}
