% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_loadLatest.R
\name{sensor_loadLatest}
\alias{sensor_loadLatest}
\title{Load hourly-aggregated Purple Air data for a week}
\usage{
sensor_loadLatest(collection = "scaqmd")
}
\arguments{
\item{collection}{Name associated with the collection.}
}
\value{
An object of class "pa_timeseries".
}
\description{
A pre-generated \code{airsensor} object will be loaded 
containing data for the most recent 7-day interval.

Each \code{airsensor} object contains data from a named collection of 
Purple Air sensors.
}
\examples{
\donttest{
setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
sensor_loadLatest("scaqmd") \%>\%
  PWFSLSmoke::monitor_timeseriesPlot(style = 'gnats')
}
}
\seealso{
\link{sensor_load}

\link{sensor_loadMonth}

\link{pat_createAirSensor}
}
