% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_videoFrame.R
\name{sensor_videoFrame}
\alias{sensor_videoFrame}
\title{Generate video frames}
\usage{
sensor_videoFrame(sensor = sensor_load(), communityRegion = NULL,
  frameTime = NULL, timeInfo = NULL, timeRange = NULL,
  timeTicks = NULL, timeLabels = NULL, map = NULL)
}
\arguments{
\item{sensor}{An AirSensor object.}

\item{communityRegion}{The area to display.}

\item{frameTime}{POSIXct specifying the time to use.}

\item{timeInfo}{???.}

\item{timeRange}{Movie time range.}

\item{timeTicks}{Movie time ticks.}

\item{timeLabels}{Movie time labels.}

\item{map}{A pre-generated basemap image.}
}
\value{
ret
}
\description{
desc
}
\examples{
ft <- lubridate::ymd_h("2019-07-04 18")
#ti <- PWFSLSmoke::timeInfo(ft, -118.082, 33.767)
sMap <- PWFSLSmoke::staticmap_getStamenmapBrick(centerLon = -118.083,
                                                centerLat = 33.767,
                                                zoom = 15,
                                                width = 770,
                                                height = 495)
sensor_videoFrame(communityRegion = "Seal Beach", 
                  frameTime = ft,
                  map = sMap)
}
