% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_videoFrame.R
\name{sensor_videoFrame}
\alias{sensor_videoFrame}
\title{Generate video frames}
\usage{
sensor_videoFrame(sensor = sensor_load(), communityRegion = NULL,
  frameTime = NULL, timeInfo = NULL, timeAxis = NULL,
  timeTicks = NULL, timeLabels = NULL, map = NULL, logo = NULL,
  colorPalette = NULL, colorBins = NULL)
}
\arguments{
\item{sensor}{An AirSensor object.}

\item{communityRegion}{The area to display.}

\item{frameTime}{POSIXct specifying the time to use.}

\item{timeInfo}{Dataframe created with \code{PWFSLSmoke::timeInfo()}.}

\item{timeAxis}{Movie time axis.}

\item{timeTicks}{Movie time ticks.}

\item{timeLabels}{Movie time labels.}

\item{map}{A pre-generated basemap image.}

\item{logo}{A PNG image to be placed near the upper right of the image.}

\item{colorPalette}{Vector of colors to use.}

\item{colorBins}{Vector of numeric breaks used to map pm25 values to colors}
}
\value{
A plot is generated. Nothing is returned.
}
\description{
Create a plot to be used as a single "frame" in a timeseries
video. This function is tailored to communities working with the South Coast
Air Quality Management District.
}
\examples{
\dontrun{
# Fourth of July in Seal Beach
library(MazamaSpatialUtils)

start <- lubridate::ymd_h("2019-07-04 00", tz = "America/Los_Angeles")
end   <- start + lubridate::days(3)

sensor <- sensor_load("scaqmd")

movieData <- 
  sensor \%>\%
  sensor_filterMeta(communityRegion == "Seal Beach") \%>\%
  sensor_filterDate(start, end)

# Seal Beach map
lon <- -118.083
lat <- 33.767
zoom <- 15

map <- PWFSLSmoke::staticmap_getStamenmapBrick(
  centerLon = lon,
  centerLat = lat,
  zoom = zoom,
  width = 770,
  height = 495
)

tickSkip <- 6
timeAxis <- movieData$data$datetime
timeTicks <- timeAxis[(lubridate::hour(timeAxis) - 1) \%\% tickSkip == 0 & 
                 lubridate::minute(timeAxis) == 0]
timeLabels <- strftime(timeTicks, "\%l \%P")
timeInfo <- PWFSLSmoke::timeInfo(timeAxis, longitude = lon, latitude = lat)

frameTime <- lubridate::ymd_h("2019-07-04 21", tz="America/Los_Angeles")
timeInfo <- PWFSLSmoke::timeInfo(frameTime, lon, lat)

sensor_videoFrame(
  sensor = sensor,
  communityRegion = "Seal Beach", 
  frameTime = frameTime,
  timeInfo = timeInfo,
  timeAxis = timeAxis,
  timeTicks = timeTicks,
  timeLabels = timeLabels,
  map = map
  # logo = png::readPNG("~/Desktop/ms_logo.png")
)
}
}
