% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor.R
\name{setArchiveBaseUrl}
\alias{setArchiveBaseUrl}
\title{Set data archive base URL}
\usage{
setArchiveBaseUrl(archiveBaseUrl)
}
\arguments{
\item{archiveBaseUrl}{Base URL pointing to an archive of pre-generated data files.}
}
\value{
Silently returns previous value of base URL.
}
\description{
Sets the package base URL pointing to an archive of
pre-generated data files.

Known base URLs include:
\itemize{
\item{http://smoke.mazamascience.com/data/PurpleAir}
\item{https://airfire-data-exports.s3-us-west-2.amazonaws.com/PurpleAir/v1}
}
}
\seealso{
ArchiveBaseUrl

getArchiveBaseUrl
}
\keyword{environment}
