% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shiny.R
\name{shiny_barplot}
\alias{shiny_barplot}
\title{Bar plot for AirShiny}
\usage{
shiny_barplot(pat, period)
}
\arguments{
\item{pat}{PurpleAir Timeseries "pat" object from \code{pat_loadLatest()}}

\item{period}{The time period to average to. Can be "sec", "min", "hour", 
"day", "DSTday", "week", "month", "quarter" or "year". A number can also
 precede these options followed by a space (i.e. "2 day" or "37 min").}
}
\description{
Bar plot for AirShiny
}
