% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoH_pctDC.R
\name{SoH_pctDC}
\alias{SoH_pctDC}
\title{Daily DC Signal percentage}
\usage{
SoH_pctDC(pat = NULL, testPeriod = "30 min")
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{testPeriod}{The period of time over which to test for a DC signal. 
Choose from "30 min", "1 hour", or "1 day".}
}
\description{
This function calculates the daily percentage of DC signal 
recorded by the \code{pm25_A}, \code{pm25_B}, \code{humidity}, and 
\code{temperature} channels. Each day is broken up into individual chunks
defined by \code{testPeriod}. The data are flagged as DC signal when the 
standard deviation of the data from each channel equals zero over a 
\code{testPeriod}. The number of chunks with a DC signal are summed over the 
day and a daily DC percentage is returned.
}
\examples{
 
tbl <- 
  example_pat_failure_A \%>\%
  SoH_pctDC(testPeriod = "30 min") 

timeseriesTbl_multiplot(tbl, ylim = c(0,100))
}
