% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoH_metrics.R
\name{SoH_pctReporting}
\alias{SoH_pctReporting}
\title{Daily reporting percentage}
\usage{
SoH_pctReporting(pat = NULL, samplingFreq = 30)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{samplingFreq}{The number of samples measured per hour when the sensor 
is operating optimally. Note that currently the sensors measure 30 samples 
per hour but were previously sampling at a higher rate.}
}
\description{
The number of sensor readings recorded per hour are summed over 
the course of a calendar day (24 hours unless a partial day is included in 
the data), then divided by the number of samples the sensor would record in 
an ideal day (30 samples/hour * 24 hours/day) to return a percentage of each 
day that the sensor is performing optimally.
}
\note{
Purple Air II sensors reporting after the June, 2019 firmware
upgrade report data every 2 minutes or 30 measurements per hour.
}
\examples{
tbl <- 
  example_pat \%>\%
  SoH_pctReporting() 

plot(tbl$datetime, tbl$pct_Reporting)

tbl <- 
  example_pat_failure_B \%>\%
  SoH_pctReporting() 

plot(tbl$datetime, tbl$pct_Reporting)


}
